refactored file structures
