--- WATCHER-test-files/WATCHER/ckl/ResultEngineInIStigSingle.ckl
@@ -1,191 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<!--<Evaluate-STIG><version>1.2310.1</version></Evaluate-STIG>-->
-
-<CHECKLIST>
-	<ASSET>
-		<ROLE>Workstation</ROLE>
-		<ASSET_TYPE>Computing</ASSET_TYPE>
-		<MARKING />
-		<HOST_NAME>test</HOST_NAME>
-		<HOST_IP></HOST_IP>
-		<HOST_MAC></HOST_MAC>
-		<HOST_FQDN></HOST_FQDN>
-		<TARGET_COMMENT></TARGET_COMMENT>
-		<TECH_AREA></TECH_AREA>
-		<TARGET_KEY></TARGET_KEY>
-		<WEB_OR_DATABASE></WEB_OR_DATABASE>
-		<WEB_DB_SITE />
-		<WEB_DB_INSTANCE />
-	</ASSET>
-	<STIGS>
-		<iSTIG>
-			<!--<Evaluate-STIG><time>2023-12-11T12:56:14.3576272-05:00</time><module><name>Scan-GoogleChrome_Checks</name><version>1.2023.7.24</version></module></Evaluate-STIG>-->
-			<STIG_INFO>
-				<SI_DATA>
-					<SID_NAME>version</SID_NAME>
-					<SID_DATA>2</SID_DATA>
-				</SI_DATA>
-				<SI_DATA>
-					<SID_NAME>classification</SID_NAME>
-					<SID_DATA>UNCLASSIFIED</SID_DATA>
-				</SI_DATA>
-				<SI_DATA>
-					<SID_NAME>customname</SID_NAME>
-				</SI_DATA>
-				<SI_DATA>
-					<SID_NAME>stigid</SID_NAME>
-					<SID_DATA>Google_Chrome_Current_Windows</SID_DATA>
-				</SI_DATA>
-				<SI_DATA>
-					<SID_NAME>description</SID_NAME>
-					<SID_DATA>This Security Technical Implementation Guide is published as a tool to improve the security of Department of Defense (DoD) information systems. The requirements are derived from the National Institute of Standards and Technology (NIST) 800-53 and related documents. Comments or proposed revisions to this document should be sent via email to the following address: disa.stig_spt@mail.mil.</SID_DATA>
-				</SI_DATA>
-				<SI_DATA>
-					<SID_NAME>filename</SID_NAME>
-					<SID_DATA>U_Google_Chrome_STIG_V2R8_Manual-xccdf.xml</SID_DATA>
-				</SI_DATA>
-				<SI_DATA>
-					<SID_NAME>releaseinfo</SID_NAME>
-					<SID_DATA>Release: 8 Benchmark Date: 26 Jan 2023</SID_DATA>
-				</SI_DATA>
-				<SI_DATA>
-					<SID_NAME>title</SID_NAME>
-					<SID_DATA>Google Chrome Current Windows Security Technical Implementation Guide</SID_DATA>
-				</SI_DATA>
-				<SI_DATA>
-					<SID_NAME>uuid</SID_NAME>
-					<SID_DATA>4055d97a-e0c8-4b20-a03c-9242f6383d93</SID_DATA>
-				</SI_DATA>
-				<SI_DATA>
-					<SID_NAME>notice</SID_NAME>
-					<SID_DATA>terms-of-use</SID_DATA>
-				</SI_DATA>
-				<SI_DATA>
-					<SID_NAME>source</SID_NAME>
-					<SID_DATA>STIG.DOD.MIL</SID_DATA>
-				</SI_DATA>
-			</STIG_INFO>
-			<VULN>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>Vuln_Num</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA>V-221558</ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>Severity</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA>medium</ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>Group_Title</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA>SRG-APP-000039</ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>Rule_ID</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA>SV-221558r769351_rule</ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>Rule_Ver</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA>DTBC-0001</ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>Rule_Title</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA>Firewall traversal from remote host must be disabled.</ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>Vuln_Discuss</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA>Remote connections should never be allowed that bypass the firewall, as there is no way to verify if they can be trusted. Enables usage of STUN and relay servers when remote clients are trying to establish a connection to this machine. If this setting is enabled, then remote clients can discover and connect to this machine even if they are separated by a firewall. If this setting is disabled and outgoing UDP connections are filtered by the firewall, then this machine will only allow connections from client machines within the local network. If this policy is left not set the setting will be enabled.	</ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>IA_Controls</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA></ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>Check_Content</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA>hi</ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>Fix_Text</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA>N/A</ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>False_Positives</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA></ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>False_Negatives</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA></ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>Documentable</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA>false</ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>Mitigations</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA></ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>Potential_Impact</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA></ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>Third_Party_Tools</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA></ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>Mitigation_Control</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA></ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>Responsibility</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA></ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>Security_Override_Guidance</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA></ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>Check_Content_Ref</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA>M</ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>Weight</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA>10.0</ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>Class</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA>Unclass</ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>STIGRef</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA>Google Chrome Current Windows Security Technical Implementation Guide :: Version 2, Release: 8 Benchmark Date: 26 Jan 2023</ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>TargetKey</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA>4081</ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>STIG_UUID</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA>6e38d515-3e60-4dcb-adf9-80623ac94a2f</ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>LEGACY_ID</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA>V-44711</ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>LEGACY_ID</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA>SV-57545</ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STIG_DATA>
-					<VULN_ATTRIBUTE>CCI_REF</VULN_ATTRIBUTE>
-					<ATTRIBUTE_DATA>CCI-001414</ATTRIBUTE_DATA>
-				</STIG_DATA>
-				<STATUS>NotAFinding</STATUS>
-				<FINDING_DETAILS>test
-</FINDING_DETAILS>
-				<COMMENTS />
-				<SEVERITY_OVERRIDE />
-				<SEVERITY_JUSTIFICATION />
-			</VULN>
-		</iSTIG>
-	
-	</STIGS>
-</CHECKLIST>
