+++ WATCHER-test-files/WATCHER/ckl/no-Stigs-xml-element.ckl
