--- WATCHER-test-files/WATCHER/cklb/MultiStig-Simple.cklb
@@ -1,3 +1,0 @@
-{
-  
-}
