--- WATCHER-test-files/WATCHER/cklb/NoASSETelement.cklb
@@ -1,3 +1,0 @@
-{
-  
-}
