--- WATCHER-test-files/WATCHER/cklb/NoResult.cklb
@@ -1,1 +1,0 @@
-{}
