--- WATCHER-test-files/WATCHER/cklb/NoRuleId.cklb
@@ -1,1 +1,0 @@
-{}
