+++ test/ckl-tests/CKLReviewParserReviewObject.test.js
@@ -1,11 +1,20 @@
 import chai from 'chai';
-import { reviewsFromCklb } from '../ReviewParser.js'; 
+import { reviewsFromCkl } from '../../ReviewParser.js'; 
+import { XMLParser } from 'fast-xml-parser';
 import fs from 'fs/promises';
-
+import he from 'he';
 
 const expect = chai.expect;
 
-
+const valueProcessor = function (
+  tagName,
+  tagValue,
+  jPath,
+  hasAttributes,
+  isLeafNode
+) {
+  he.decode(tagValue)
+}
 
 // Create a helper function to read the file and generate the review object
 async function generateReviewObject (
@@ -15,26 +24,28 @@ async function generateReviewObject (
   allowAccept
 ) {
   const data = await fs.readFile(filePath, 'utf8')
-  return reviewsFromCklb({
+  return reviewsFromCkl({
     data,
+    importOptions,
     fieldSettings,
     allowAccept,
-    importOptions,
+    valueProcessor,
+    XMLParser
   })
 }
 
 
-describe('Import Options, allowAccept for a CKLb review object in non multi-stig', () => {
+describe('Import Options, allowAccept for a CKL review object in non multi-stig', () => {
   it('DEFAULT SETTINGS: Primarily testing review "status = saved"', async () => {
     // Test: DEFAULT SETTINGS
-    // This test validates the behavior of the cklb parser function under default settings.
+    // This test validates the behavior of the ckl parser function under default settings.
     // Primary Focus:
     // - Ensuring that the 'autoStatus' option in 'importOptions' correctly sets the review status to 'saved'.
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // These checks ensure that not only is the 'status' property set as expected, but also that
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-notReviewed-Commented-Detailed.cklb') to simulate a real-world scenario.
+    // The test utilizes a sample CKL file ('Single-Vuln-notReviewed-Commented-Detailed.ckl') to simulate a real-world scenario.
 
     const importOptions = {
       autoStatus: 'saved',
@@ -59,7 +70,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-notReviewed-Commented-Detailed.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-notReviewed-Commented-Detailed.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -67,10 +78,9 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
       fieldSettings,
       allowAccept
     )
-    
 
     const expectedReview = {
-      ruleId: 'SV-257777r925317_rule',
+      ruleId: 'SV-257777r925318_rule',
       status: 'saved',
       result: 'informational',
       comment: 'xyz',
@@ -82,13 +92,13 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
 
   it('autoStatus = null (keep exisiting), testing that all review statuses do not exist', async () => {
     // Test: autoStatus = null
-    // This test validates the behavior of the cklb parser function under the 'autoStatus = null' setting.
+    // This test validates the behavior of the ckl parser function under the 'autoStatus = null' setting.
     // Primary Focus:
     // - Ensuring that the 'autoStatus' option in 'importOptions' correctly sets the review status to null (doesnt exist).
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'ruleId', 'result', 'comment', and 'detail'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-notReviewed-Commented-Detailed.cklb') to simulate a real-world scenario.
+    // The test utilizes a sample CKL file ('Single-Vuln-notReviewed-Commented-Detailed.ckl') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'null',
       unreviewed: 'commented',
@@ -112,7 +122,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-notReviewed-Commented-Detailed.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-notReviewed-Commented-Detailed.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -122,7 +132,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     )
 
     const expectedReview = {
-      ruleId: 'SV-257777r925317_rule',
+      ruleId: 'SV-257777r925318_rule',
       result: 'informational',
       comment: 'xyz',
       detail: 'xyz'
@@ -135,14 +145,14 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
 
   it('autoStatus = submitted, testing if reviews are set to "submitted" if valid or "saved" if not valid. Determined by field settings and result', async () => {
     // Test: autoStatus = submitted
-    // This test validates the behavior of the cklb parser function under the autoStatus = submitted setting coupled with out field settings.
+    // This test validates the behavior of the ckl parser function under the autoStatus = submitted setting coupled with out field settings.
     // Primary Focus:
     // - Ensuring that the 'autoStatus' option in 'importOptions' correctly sets the review status to 'submitted' if valid or 'saved' if not valid.
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including ruleId 'result', 'comment', and 'detail'.
     // These checks ensure that not only is the 'status' property set as expected, but also that
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-notReviewed-Commented-Detailed.cklb') to simulate a real-world scenario.
+    // The test utilizes a sample CKL file ('Single-Vuln-notReviewed-Commented-Detailed.ckl') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -166,7 +176,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Asset_a-VPN_TRUNCATED-V2R5.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Asset_a-VPN_TRUNCATED-V2R5.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -209,8 +219,8 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     }
 
     // ensuring that each review has a status that matches the expected status
-    for (const checklists of review.checklists) {
-      for (const reviewItem of checklists.reviews) {
+    for (const checklist of review.checklists) {
+      for (const reviewItem of checklist.reviews) {
         const expectedStatus = expectedStatuses[reviewItem.ruleId]
         const expectedComment = expectedComments[reviewItem.ruleId]
         const expectedDetail = expectedDetails[reviewItem.ruleId]
@@ -223,14 +233,14 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
 
   it('autoStatus = submitted, testing that we will set status to saved if does not meet field settings requirements', async () => {
     // Test: autoStatus = submitted
-    // This test validates the behavior of the cklb parser function under the autoStatus = submitted setting coupled with our field settings.
+    // This test validates the behavior of the ckl parser function under the autoStatus = submitted setting coupled with our field settings.
     // Primary Focus:
     // - Ensuring that the 'autoStatus' option in 'importOptions' correctly sets the review status to 'saved' if not valid.
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail'.
     // These checks ensure that not only is the 'status' property set as expected, but also that
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-notReviewed-Commented-Detailed.cklb') to simulate a real-world scenario.
+    // The test utilizes a sample CKL file ('Single-Vuln-notReviewed-Commented-Detailed.ckl') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -254,7 +264,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-notReviewed-Commented-Detailed.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-notReviewed-Commented-Detailed.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -263,7 +273,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
       allowAccept
     )
     const expectedReview = {
-      ruleId: 'SV-257777r925317_rule',
+      ruleId: 'SV-257777r925318_rule',
       status: 'saved',
       result: 'informational',
       comment: 'xyz',
@@ -274,7 +284,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
 
   it('(autoStatus = accepted, allowAccept = true) for permissions to use accepted status', async () => {
     // Test: autoStatus = accepted, allowAccept = true
-    // This test validates the behavior of the cklb parser function under the autoStatus = accepted, allowAccept = true settings.
+    // This test validates the behavior of the ckl parser function under the autoStatus = accepted, allowAccept = true settings.
     // Primary Focus:
     // - Ensuring that the 'autoStatus' option in 'importOptions' correctly sets the review status to 'submitted' if the user does not have permissions
     // or 'accepted' if the user does have permissions (permissions are determined by the 'allowAccept' option).
@@ -284,7 +294,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // These checks ensure that not only is the 'status' property set as expected, but also that
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Asset_a-VPN_TRUNCATED-V2R5.cklb'') to simulate a real-world scenario.
+    // The test utilizes a sample CKL file ('Asset_a-VPN_TRUNCATED-V2R5.ckl'') to simulate a real-world scenario.
 
     const importOptions = {
       autoStatus: 'accepted',
@@ -308,7 +318,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
 
     const allowAccept = true
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Asset_a-VPN_TRUNCATED-V2R5.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Asset_a-VPN_TRUNCATED-V2R5.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -365,10 +375,9 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
       expect(reviewItem.detail).to.equal(expectedDetail)
     }
   })
-
   it('(autoStatus = accepted allowAccept = false) for permissions to use submitted status', async () => {
     // Test: autoStatus = accepted, allowAccept = false
-    // This test validates the behavior of the cklb parser function under the autoStatus = accepted, allowAccept = false settings.
+    // This test validates the behavior of the ckl parser function under the autoStatus = accepted, allowAccept = false settings.
     // Primary Focus:
     // - Ensuring that the 'autoStatus' option in 'importOptions' correctly sets the review status to 'submitted' if the user does not have permissions
     // or 'accepted' if the user does have permissions (permissions are determined by the 'allowAccept' option).
@@ -378,7 +387,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // These checks ensure that not only is the 'status' property set as expected, but also that
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Asset_a-VPN_TRUNCATED-V2R5.cklb'') to simulate a real-world scenario.
+    // The test utilizes a sample CKL file ('Asset_a-VPN_TRUNCATED-V2R5.ckl'') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'accepted',
       unreviewed: 'commented',
@@ -401,7 +410,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
 
     const allowAccept = false
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Asset_a-VPN_TRUNCATED-V2R5.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Asset_a-VPN_TRUNCATED-V2R5.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -460,14 +469,14 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
 
   it("'unreviewed = commented' testing that we only import unreviwred rules that contain a comment or detail   ", async () => {
     // Test: unreviewed = commented
-    // This test validates the behavior of the cklb parser function under the unreviewed = commented settings.
+    // This test validates the behavior of the ckl parser function under the unreviewed = commented settings.
     // Primary Focus:
     // - Ensuring that the 'unreviewed' option in 'importOptions' correctly imports only reviews with a non complience result that contain a comment and or detail.
     // note this test will have a review with a comment and detail
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-notReviewed-Commented-Detailed.cklb') to simulate a real-world scenario.
+    // The test utilizes a sample CKL file ('Single-Vuln-notReviewed-Commented-Detailed.ckl') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -491,7 +500,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-notReviewed-Commented-Detailed.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-notReviewed-Commented-Detailed.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -501,7 +510,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     )
 
     const expectedReview = {
-      ruleId: 'SV-257777r925317_rule',
+      ruleId: 'SV-257777r925318_rule',
       status: 'saved',
       result: 'informational',
       comment: 'xyz',
@@ -519,14 +528,14 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
 
   it("'unreviewed = commented' testing that we only import unreviwred rules that contain a comment or detail but giving a review without either", async () => {
     // Test: unreviewed = commented
-    // This test validates the behavior of the cklb parser function under the unreviewed = commented settings.
+    // This test validates the behavior of the ckl parser function under the unreviewed = commented settings.
     // Primary Focus:
     // - Ensuring that the 'unreviewed' option in 'importOptions' correctly imports only reviews with a non complience result that contain a comment and or detail.
     // note this test will not have any reviews because there will be no comment/detail
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-notReviewed-Commented-Detailed.cklb') to simulate a real-world scenario.
+    // The test utilizes a sample CKL file ('Single-Vuln-notReviewed-Commented-Detailed.ckl') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -550,7 +559,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-notReviewed-Empty-CommentDetail.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-notReviewed-Empty-CommentDetail.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -558,21 +567,21 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
       fieldSettings,
       allowAccept
     )
-
+      
     // check that review has a comment or detail and exisits
     expect(review.checklists[0].reviews).to.be.empty
   })
 
   it("'unreviewed = always', testing that unreviewed always are import  ", async () => {
     // Test: unreviewed = always
-    // This test validates the behavior of the cklb parser function under the unreviewed = always settings.
+    // This test validates the behavior of the ckl parser function under the unreviewed = always settings.
     // Primary Focus:
     // - Ensuring that the 'unreviewed' option in 'importOptions' correctly imports all reviews without a compliance result and if they have a comment
     // or detail they will be labeled as informational.
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-notReviewed-Commented-Detailed.cklb') to simulate a real-world scenario.
+    // The test utilizes a sample CKL file ('Single-Vuln-notReviewed-Commented-Detailed.ckl') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'always',
@@ -596,7 +605,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-notReviewed-Commented-Detailed.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-notReviewed-Commented-Detailed.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -606,7 +615,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     )
 
     const expectedReview = {
-      ruleId: 'SV-257777r925317_rule',
+      ruleId: 'SV-257777r925318_rule',
       status: 'saved',
       result: 'informational',
       comment: 'xyz',
@@ -615,17 +624,16 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
 
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
-
   it("'unreviewed = always', testing that unreviewed always are import  ", async () => {
     // Test: unreviewed = always
-    // This test validates the behavior of the cklb parser function under the unreviewed = always settings.
+    // This test validates the behavior of the ckl parser function under the unreviewed = always settings.
     // Primary Focus:
     // - Ensuring that the 'unreviewed' option in 'importOptions' correctly imports all reviews without a compliance result
     // and without comment or detail to be labled as notchecked.
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // These checks ensure that not only is the 'status' property set as expected, but also that
-    // The test utilizes a sample CKLb file ('Single-Vuln-notReviewed-Empty-CommentDetail.cklb') to simulate a real-world scenario.
+    // The test utilizes a sample CKL file ('Single-Vuln-notReviewed-Empty-CommentDetail.ckl') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'always',
@@ -649,7 +657,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-notReviewed-Empty-CommentDetail.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-notReviewed-Empty-CommentDetail.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -657,9 +665,8 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
       fieldSettings,
       allowAccept
     )
-   
     const expectedReview = {
-      ruleId: 'SV-257777r925317_rule',
+      ruleId: 'SV-257777r925318_rule',
       status: 'saved',
       result: 'notchecked',
       comment: null,
@@ -668,16 +675,15 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
 
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
-
   it(" 'unreviewed = never' testing to never import an unreviewed item ", async () => {
     // Test: unreviewed = never
-    // This test validates the behavior of the cklb parser function under the unreviewed = never settings.
+    // This test validates the behavior of the ckl parser function under the unreviewed = never settings.
     // Primary Focus:
     // - Ensuring that the 'unreviewed' option in 'importOptions' ignores reviews without a compliance result (Nf/na/o)
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-notReviewed-Empty-CommentDetail.cklb') to simulate a real-world scenario.
+    // The test utilizes a sample CKL file ('Single-Vuln-notReviewed-Empty-CommentDetail.ckl') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'never',
@@ -701,7 +707,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-notReviewed-Empty-CommentDetail.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-notReviewed-Empty-CommentDetail.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -715,14 +721,14 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
 
   it("'unreviewedComment = informational' testing that an unreviewed item with a comment has a result of informational", async () => {
     // Test: unreviewedComment = informational
-    // This test validates the behavior of the cklb parser function under the unreviewedComment = informational settings.
+    // This test validates the behavior of the ckl parser function under the unreviewedComment = informational settings.
     // Primary Focus:
     // - Ensuring that the 'unreviewed' option in 'importOptions' correctly labels a review without a compliance result..
     // and with a comment or detail as informational.
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-notReviewed-Commented-Detailed.cklb') to simulate a real-world scenario.
+    // The test utilizes a sample CKL file ('Single-Vuln-notReviewed-Commented-Detailed.ckl') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -745,7 +751,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-notReviewed-Commented-Detailed.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-notReviewed-Commented-Detailed.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -755,7 +761,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     )
 
     const expectedReview = {
-      ruleId: 'SV-257777r925317_rule',
+      ruleId: 'SV-257777r925318_rule',
       status: 'saved',
       result: 'informational',
       comment: 'xyz',
@@ -766,14 +772,14 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
   })
   it(" 'unreviewedComment = notchecked'. testing that an unreviewed with a comment has a result of notchecked", async () => {
     // Test: unreviewedComment = notchecked
-    // This test validates the behavior of the cklb parser function under the unreviewedComment = notchecked settings.
+    // This test validates the behavior of the ckl parser function under the unreviewedComment = notchecked settings.
     // Primary Focus:
     // - Ensuring that the 'unreviewed' option in 'importOptions' correctly labels a review without a compliance result..
     // and with a comment or detail as notchecked.
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-notReviewed-Commented-Detailed.cklb') to simulate a real-world scenario.
+    // The test utilizes a sample CKL file ('Single-Vuln-notReviewed-Commented-Detailed.ckl') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -796,7 +802,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-notReviewed-Commented-Detailed.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-notReviewed-Commented-Detailed.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -805,7 +811,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
       allowAccept
     )
     const expectedReview = {
-      ruleId: 'SV-257777r925317_rule',
+      ruleId: 'SV-257777r925318_rule',
       status: 'saved',
       result: 'notchecked',
       comment: 'xyz',
@@ -814,16 +820,15 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
 
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
-
   it("'emptyDetail = replace' testing that if an item has an empty detail we will replace it with a static message", async () => {
     // Test: emptyDetail = replace
-    // This test validates the behavior of the cklb parser function under the emptyDetail = replace settings.
+    // This test validates the behavior of the ckl parser function under the emptyDetail = replace settings.
     // Primary Focus:
     // - Ensuring that the 'emptyDetail' option in 'importOptions' correctly replaces an empty detail with a static message. "There is no detail provided for the assessment"
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-Pass-With-Comment.cklb') to simulate a real-world scenario.
+    // The test utilizes a sample CKL file ('Single-Vuln-Pass-With-Comment.ckl') to simulate a real-world scenario.
 
     const importOptions = {
       autoStatus: 'saved',
@@ -847,7 +852,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-Pass-With-Comment.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-Pass-With-Comment.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -856,24 +861,23 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
       allowAccept
     )
     const expectedReview = {
-      ruleId: 'SV-207191r803417_rule',
+      ruleId: 'SV-207191r803418_rule',
       status: 'saved',
       result: 'pass',
       comment: 'xyz',
       detail: 'There is no detail provided for the assessment'
     }
-
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
   it("'emptyDetail = ignore' testing that if there is no detail provided it will retaing exisiting, if no exisitng then we will set to null", async () => {
     // Test: emptyDetail = ignore
-    // This test validates the behavior of the cklb parser function under the emptyDetail = ignore settings.
+    // This test validates the behavior of the ckl parser function under the emptyDetail = ignore settings.
     // Primary Focus:
     // - Ensuring that the 'emptyDetail' option in 'importOptions' correctly replaces an empty detail a null value
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-Pass-With-Comment.cklb') to simulate a real-world scenario.
+    // The test utilizes a sample CKL file ('Single-Vuln-Pass-With-Comment.ckl') to simulate a real-world scenario.
 
     const importOptions = {
       autoStatus: 'saved',
@@ -897,7 +901,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-Pass-With-Comment.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-Pass-With-Comment.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -907,7 +911,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803417_rule',
+      ruleId: 'SV-207191r803418_rule',
       status: 'saved',
       result: 'pass',
       comment: 'xyz',
@@ -918,13 +922,13 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
   })
   it("'emptyDetail = import', testing empty detail will clear existing text (setting it to empty string)", async () => {
     // Test: emptyDetail = import
-    // This test validates the behavior of the cklb parser function under the emptyDetail = import settings.
+    // This test validates the behavior of the ckl parser function under the emptyDetail = import settings.
     // Primary Focus:
     // - Ensuring that the 'emptyDetail' option in 'importOptions' correctly replaces an empty detail an empty string if no detail is provided
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-Pass-With-Comment.cklb') to simulate a real-world scenario.
+    // The test utilizes a sample CKL file ('Single-Vuln-Pass-With-Comment.ckl') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -947,7 +951,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-Pass-With-Comment.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-Pass-With-Comment.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -957,7 +961,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803417_rule',
+      ruleId: 'SV-207191r803418_rule',
       status: 'saved',
       result: 'pass',
       comment: 'xyz',
@@ -969,13 +973,13 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
 
   it(" 'emptyDetail = import' testing that a review with a detail provided will be applied", async () => {
     // Test: emptyDetail = import
-    // This test validates the behavior of the cklb parser function under the emptyDetail = import settings.
+    // This test validates the behavior of the ckl parser function under the emptyDetail = import settings.
     // Primary Focus:
     // - Ensuring that the 'emptyDetail' option in 'importOptions' correctly uses the exisisitng detail if one is provided
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-Pass-With-Detail.cklb') to simulate a real-world scenario.
+    // The test utilizes a sample CKL file ('Single-Vuln-Pass-With-Detail.ckl') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -998,7 +1002,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-Pass-With-Detail.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-Pass-With-Detail.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -1008,7 +1012,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803417_rule',
+      ruleId: 'SV-207191r803418_rule',
       status: 'saved',
       result: 'pass',
       comment: null,
@@ -1019,13 +1023,13 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
 
   it("'emptyComment = replace' testing that if an item has an empty comment we will replace it with a static message", async () => {
     // Test: emptyComment = replace
-    // This test validates the behavior of the cklb parser function under the emptyComment = replace settings.
+    // This test validates the behavior of the ckl parser function under the emptyComment = replace settings.
     // Primary Focus:
     // - Ensuring that the 'emptyComment' option in 'importOptions' correctly replaces an empty comment with a static message. "There is no comment provided for the assessment"
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-Pass-With-Detail.cklb') to simulate a real-world scenario.
+    // The test utilizes a sample CKL file ('Single-Vuln-Pass-With-Detail.ckl') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -1048,7 +1052,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-Pass-With-Detail.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-Pass-With-Detail.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -1058,7 +1062,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803417_rule',
+      ruleId: 'SV-207191r803418_rule',
       status: 'saved',
       result: 'pass',
       comment: 'There is no comment provided for the assessment',
@@ -1067,16 +1071,15 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
 
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
-
   it("'emptyComment = ignore' testing that we will use exisitng text, if none use null", async () => {
     // Test: emptyComment = ignore
-    // This test validates the behavior of the cklb parser function under the emptyComment = ignore settings.
+    // This test validates the behavior of the ckl parser function under the emptyComment = ignore settings.
     // Primary Focus:
     // - Ensuring that the 'emptyComment' option in 'importOptions' correctly replaces an empty comment a null value
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-Pass-With-Comment.cklb') to simulate a real-world scenario.
+    // The test utilizes a sample CKL file ('Single-Vuln-Pass-With-Comment.ckl') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -1099,7 +1102,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-Pass-With-Detail.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-Pass-With-Detail.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -1108,7 +1111,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
       allowAccept
     )
     const expectedReview = {
-      ruleId: 'SV-207191r803417_rule',
+      ruleId: 'SV-207191r803418_rule',
       status: 'saved',
       result: 'pass',
       comment: null,
@@ -1116,16 +1119,15 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     }
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
-
   it("'emptyComment = import', will clear eixsitng with an empty string if no comment given ", async () => {
     // Test: emptyComment = import
-    // This test validates the behavior of the cklb parser function under the emptyComment = import settings.
+    // This test validates the behavior of the ckl parser function under the emptyComment = import settings.
     // Primary Focus:
     // - Ensuring that the 'emptyComment' option in 'importOptions' correctly replaces an empty comment an empty string: ""
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-Pass-With-Detail.cklb')
+    // The test utilizes a sample CKL file ('Single-Vuln-Pass-With-Detail.ckl')
 
     const importOptions = {
       autoStatus: 'saved',
@@ -1149,7 +1151,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-Pass-With-Detail.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-Pass-With-Detail.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -1159,7 +1161,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803417_rule',
+      ruleId: 'SV-207191r803418_rule',
       status: 'saved',
       result: 'pass',
       comment: '',
@@ -1167,16 +1169,15 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     }
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
-
-  it("'emptyComment = import', testing a review with a comment provided in the cklb to make sure we get it back in the review ", async () => {
+  it("'emptyComment = import', testing a review with a comment provided in the ckl to make sure we get it back in the review ", async () => {
     // Test: emptyComment = import
-    // This test validates the behavior of the cklb parser function under the emptyComment = import settings.
+    // This test validates the behavior of the ckl parser function under the emptyComment = import settings.
     // Primary Focus:
     // - Ensuring that the 'emptyComment' option in 'importOptions' correctly uses the exisisitng comment if one is provided
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-Pass-With-Detail.cklb')
+    // The test utilizes a sample CKL file ('Single-Vuln-Pass-With-Detail.ckl')
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -1199,7 +1200,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-Pass-With-Comment.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-Pass-With-Comment.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -1209,7 +1210,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803417_rule',
+      ruleId: 'SV-207191r803418_rule',
       status: 'saved',
       result: 'pass',
       comment: 'xyz',
@@ -1219,16 +1220,16 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
   })
 })
 
- describe('fieldSettings testing for a CKLb review object in non multi-stig', () => {
+describe('fieldSettings testing for a CKL review object in non multi-stig', () => {
   it("DEFAULT FIELD SETTINGS with allowAccept=true and a passing review, testing that it has a detail and is 'submitted'", async () => {
     // Test: autostatus = submitted, default field settings.
-    // This test validates the behavior of the cklb parser function under the default field settings.
+    // This test validates the behavior of the ckl parser function under the default field settings.
     // Primary Focus:
     // - Ensuring that the 'autoStatus' option in 'importOptions' correctly ensures that a detail is required for a review to be submitted.
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-Pass-With-Detail.cklb')
+    // The test utilizes a sample CKL file ('Single-Vuln-Pass-With-Detail.ckl')
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1252,7 +1253,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-Pass-With-Detail.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-Pass-With-Detail.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -1262,7 +1263,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803417_rule',
+      ruleId: 'SV-207191r803418_rule',
       status: 'submitted',
       result: 'pass',
       comment: '',
@@ -1274,14 +1275,14 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
   })
   it('DEFAULT FIELD SETTINGS  with allowAccept=true and a failing review with no detail.', async () => {
     //autostatus = submitted, testing default field settings with allowAccept=true
-    // This test validates the behavior of the cklb parser function under the default field setting with a fail and a no detail.
+    // This test validates the behavior of the ckl parser function under the default field setting with a fail and a no detail.
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a detail is always required for submission for a review.
     // Test that with a failing review and no detail exisiting and it will be set to 'saved'
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-fail-With-Detail.cklb')
+    // The test utilizes a sample CKL file ('Single-Vuln-fail-With-Detail.ckl')
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1305,7 +1306,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-fail-Empty-CommentDetail.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-fail-Empty-CommentDetail.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -1315,7 +1316,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803417_rule',
+      ruleId: 'SV-207191r803418_rule',
       status: 'saved',
       result: 'fail',
       comment: null,
@@ -1326,14 +1327,14 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
   })
   it('fieldSettings.detail.required = findings with allowAccept=true with a failing review containing a detail', async () => {
     // Test: autostatus = submitted, default field settings.
-    // This test validates the behavior of the cklb parser function under the  fieldSettings.detail.required = findings field settings.
+    // This test validates the behavior of the ckl parser function under the  fieldSettings.detail.required = findings field settings.
     // Primary Focus:
     // - Ensuring that the 'autoStatus' option in 'importOptions' correctly ensures that a detail is required for a review  that has findings to be submitted.
     // Test that a failing review with a detail will be submitted
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-Pass-With-Detail.cklb')
+    // The test utilizes a sample CKL file ('Single-Vuln-Pass-With-Detail.ckl')
 
     const importOptions = {
       autoStatus: 'submitted',
@@ -1358,7 +1359,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-fail-With-Detail.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-fail-With-Detail.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -1368,7 +1369,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803417_rule',
+      ruleId: 'SV-207191r803418_rule',
       status: 'submitted',
       result: 'fail',
       comment: '',
@@ -1379,14 +1380,14 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
   })
   it("'fieldSettings.detail.required = findings' with allowAccept=true with a fail and no detail or comment", async () => {
     // Test: autostatus = submitted, testing 'fieldSettings.detail.required = findings' allowAccept=true
-    // This test validates the behavior of the cklb parser function under the  fieldSettings.detail.required = findings field setting with a fail and no detail or comment.
+    // This test validates the behavior of the ckl parser function under the  fieldSettings.detail.required = findings field setting with a fail and no detail or comment.
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a detail is required for a review that has findings to be submitted if not we will save.
     // Test that no detail exisitng and it will be set to 'saved'
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-fail-Empty-CommentDetail.cklb')
+    // The test utilizes a sample CKL file ('Single-Vuln-fail-Empty-CommentDetail.ckl')
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1410,7 +1411,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-fail-Empty-CommentDetail.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-fail-Empty-CommentDetail.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -1419,7 +1420,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
       allowAccept
     )
     const expectedReview = {
-      ruleId: 'SV-207191r803417_rule',
+      ruleId: 'SV-207191r803418_rule',
       status: 'saved',
       result: 'fail',
       comment: null,
@@ -1433,14 +1434,14 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
 
   it("'fieldSettings.detail.required = optional' with allowAccept=true with a fail and no detail or comment, testing that it does not have a detail and is submitted ", async () => {
     // TEST: autostatus = submitted, testing 'fieldSettings.detail.required = optional' with allowAccept=true
-    // This test validates the behavior of the cklb parser function under the  fieldSettings.detail.required = optional field setting with a fail and no detail or comment.
+    // This test validates the behavior of the ckl parser function under the  fieldSettings.detail.required = optional field setting with a fail and no detail or comment.
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a detail is optional for submission for a review that has findings.
     // Test that no detail exisitng and it will be set to 'submitted'
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-fail-Empty-CommentDetail.cklb')
+    // The test utilizes a sample CKL file ('Single-Vuln-fail-Empty-CommentDetail.ckl')
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1464,7 +1465,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-fail-Empty-CommentDetail.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-fail-Empty-CommentDetail.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -1474,7 +1475,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803417_rule',
+      ruleId: 'SV-207191r803418_rule',
       status: 'submitted',
       result: 'fail',
       comment: null,
@@ -1485,14 +1486,14 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
   })
   it("fieldSettings.detail.required = optional' with allowAccept=true with a fail and detail testing it has a detail and is submitted", async () => {
     // autostatus = submitted, 'fieldSettings.detail.required = optional' with allowAccept=true
-    // This test validates the behavior of the cklb parser function under the  fieldSettings.detail.required = optional field setting with a fail and a detail.
+    // This test validates the behavior of the ckl parser function under the  fieldSettings.detail.required = optional field setting with a fail and a detail.
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a detail is optional for submission for a review that has findings.
     // Test that with detail exisitng and it will be set to 'submitted'
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-fail-With-Detail.cklb')
+    // The test utilizes a sample CKL file ('Single-Vuln-fail-With-Detail.ckl')
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1516,7 +1517,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-fail-With-Detail.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-fail-With-Detail.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -1526,7 +1527,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803417_rule',
+      ruleId: 'SV-207191r803418_rule',
       status: 'submitted',
       result: 'fail',
       comment: null,
@@ -1538,14 +1539,14 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
 
   it('DEFAULT FIELD SETTINGS with allowAccept=true and a passing review testing that it has a comment and is submitted', async () => {
     //autostatus = submitted, testing default field settings with allowAccept=true
-    // This test validates the behavior of the cklb parser function under the default field setting with a pass and a comment.
+    // This test validates the behavior of the ckl parser function under the default field setting with a pass and a comment.
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a comment is always required for submission for a review.
     // Test that witha passing review and a comment exisiting and it will be set to 'submitted'
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-fail-With-Detail.cklb')
+    // The test utilizes a sample CKL file ('Single-Vuln-fail-With-Detail.ckl')
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1569,7 +1570,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-Pass-With-Comment.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-Pass-With-Comment.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -1579,7 +1580,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803417_rule',
+      ruleId: 'SV-207191r803418_rule',
       status: 'submitted',
       result: 'pass',
       comment: 'xyz',
@@ -1590,14 +1591,14 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
   })
   it('DEFAULT FIELD SETTINGS  with allowAccept=true and a failing review with no comment.', async () => {
     //autostatus = submitted, testing default field settings with allowAccept=true
-    // This test validates the behavior of the cklb parser function under the default field setting with a fail and a no comment.
+    // This test validates the behavior of the ckl parser function under the default field setting with a fail and a no comment.
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a comment is always required for submission for a review.
     // Test that with a failing review and no comment exisiting and it will be set to 'saved'
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-fail-With-Detail.cklb')
+    // The test utilizes a sample CKL file ('Single-Vuln-fail-With-Detail.ckl')
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1621,7 +1622,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-fail-Empty-CommentDetail.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-fail-Empty-CommentDetail.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -1631,7 +1632,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803417_rule',
+      ruleId: 'SV-207191r803418_rule',
       status: 'saved',
       result: 'fail',
       comment: null,
@@ -1642,14 +1643,14 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
   })
   it('fieldSettings.comment.required = findings with allowAccept=true with a fail and comment testing that it has a comment and is submitted', async () => {
     // TEST: autostatus = submitted, fieldSettings.comment.required = findings with allowAccept=true
-    // This test validates the behavior of the cklb parser function under the fieldSettings.comment.required = findings field setting with a fail and a comment.
+    // This test validates the behavior of the ckl parser function under the fieldSettings.comment.required = findings field setting with a fail and a comment.
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a comment is always required for submission for a review that contains a finding.
     // Test that with a failed review and a comment exisitng and it will be set to 'submitted'
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-fail-with-Comment.cklb')
+    // The test utilizes a sample CKL file ('Single-Vuln-fail-with-Comment.ckl')
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1673,7 +1674,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-fail-with-Comment.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-fail-with-Comment.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -1683,7 +1684,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803417_rule',
+      ruleId: 'SV-207191r803418_rule',
       status: 'submitted',
       result: 'fail',
       comment: 'xyz',
@@ -1694,14 +1695,14 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
   })
   it("'fieldSettings.comment.required = findings' with allowAccept=true with a fail and no detail or comment, testing that it does not have a comment and is 'saved' ", async () => {
     // TEST: autostatus = submitted, testing 'fieldSettings.comment.required = findings' with allowAccept=true
-    // This test validates the behavior of the cklb parser function under the fieldSettings.comment.required = findings field setting with a fail and no comment.
+    // This test validates the behavior of the ckl parser function under the fieldSettings.comment.required = findings field setting with a fail and no comment.
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a comment is always required for submission for a review that contains a finding.
     // Test that with a failed review and no comment will be set to 'saved'
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-fail-Empty-CommentDetail.cklb')
+    // The test utilizes a sample CKL file ('Single-Vuln-fail-Empty-CommentDetail.ckl')
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1725,7 +1726,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-fail-Empty-CommentDetail.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-fail-Empty-CommentDetail.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -1735,7 +1736,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803417_rule',
+      ruleId: 'SV-207191r803418_rule',
       status: 'saved',
       result: 'fail',
       comment: null,
@@ -1747,14 +1748,14 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
 
   it("'fieldSettings.comment.required = optional' with allowAccept=true with a fail and no detail or comment. testing that it doesnt have a comment and is submmited", async () => {
     // TEST: autostatus = submitted, testing 'fieldSettings.comment.required = optional' with allowAccept=true
-    // This test validates the behavior of the cklb parser function under the fieldSettings.comment.required = optional field setting with a fail and no comment or detail .
+    // This test validates the behavior of the ckl parser function under the fieldSettings.comment.required = optional field setting with a fail and no comment or detail .
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a comment is optionally required for submission for a review that contains a finding.
     // Test that with a failed review and no comment will be set to 'submitted'
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-fail-Empty-CommentDetail.cklb')
+    // The test utilizes a sample CKL file ('Single-Vuln-fail-Empty-CommentDetail.ckl')
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1778,7 +1779,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-fail-Empty-CommentDetail.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-fail-Empty-CommentDetail.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -1788,7 +1789,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803417_rule',
+      ruleId: 'SV-207191r803418_rule',
       status: 'submitted',
       result: 'fail',
       comment: null,
@@ -1799,14 +1800,14 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
   })
   it("fieldSettings.comment.required = optional' with allowAccept=true with a fail and comment, testing thhat it has a comment and is submitted", async () => {
     // TEST: autostatus = submitted, 'fieldSettings.comment.required = optional' with allowAccept=true
-    // This test validates the behavior of the cklb parser function under the fieldSettings.comment.required = optional field setting with a fail and a comment.
+    // This test validates the behavior of the ckl parser function under the fieldSettings.comment.required = optional field setting with a fail and a comment.
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a comment is optionally required for submission for a review that contains a finding.
     // Test that with a failed review and a comment will be set to 'submitted'
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-fail-with-Comment.cklb')
+    // The test utilizes a sample CKL file ('Single-Vuln-fail-with-Comment.ckl')
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1830,7 +1831,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-fail-with-Comment.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-fail-with-Comment.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -1840,7 +1841,7 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803417_rule',
+      ruleId: 'SV-207191r803418_rule',
       status: 'submitted',
       result: 'fail',
       comment: 'xyz',
@@ -1849,19 +1850,19 @@ describe('Import Options, allowAccept for a CKLb review object in non multi-stig
 
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
- })
+})
 
-describe(' CKLb Tests where fieldSettings and importOptions overlap. ', () => {
+describe(' CKL Tests where fieldSettings and importOptions overlap. ', () => {
   it("Testing where emptyDetail: 'ignore', emptyComment: 'ignore', aswell as requiring a comment and detail  ", async () => {
     // TEST: emptyDetail: 'ignore', emptyComment: 'ignore', fieldSettings.detail.required = always, fieldSettings.comment.required = always
-    // This test validates the behavior of the cklb parser function under above settings with a non compliance resilt and no comment or detail.
+    // This test validates the behavior of the ckl parser function under above settings with a non compliance resilt and no comment or detail.
     // Primary Focus:
     // - ensuring that we will have no reviews because a brand new review will be created with "null" comment or detail with are both required.
     // Test that with a failed review and a comment will be set to 'submitted'
     // Secondary Focus:
     // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
     // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKLb file ('Single-Vuln-fail-with-Comment.cklb')
+    // The test utilizes a sample CKL file ('Single-Vuln-fail-with-Comment.ckl')
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1885,7 +1886,7 @@ describe(' CKLb Tests where fieldSettings and importOptions overlap. ', () => {
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-notReviewed-Empty-CommentDetail.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-notReviewed-Empty-CommentDetail.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -1898,7 +1899,7 @@ describe(' CKLb Tests where fieldSettings and importOptions overlap. ', () => {
   })
 })
 
-describe('MISC CKLb. ', () => {
+describe('MISC CKL. ', () => {
   it('Testing that long comment.detail is truncated ', async () => {
     // NOTE: the input comment and detail are '32768' characters long
 
@@ -1925,7 +1926,7 @@ describe('MISC CKLb. ', () => {
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-Long-CommentDetail.cklb'
+      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-Long-CommentDetail.ckl'
 
     const review = await generateReviewObject(
       filePath,
@@ -1939,5 +1940,76 @@ describe('MISC CKLb. ', () => {
     expect(review.checklists[0].reviews[0].detail).to.have.lengthOf(maxLength)
     expect(review.checklists[0].reviews[0].comment).to.have.lengthOf(maxLength)
   })
-  
+ 
+  it('Testing no <STATUS> value in a ckl ', async () => {
+    // NOTE: the input comment and detail are '32768' characters long
+
+    const importOptions = {
+      autoStatus: 'submitted',
+      unreviewed: 'commented',
+      unreviewedCommented: 'informational',
+      emptyDetail: 'ignore',
+      emptyComment: 'ignore',
+      allowCustom: true
+    }
+
+    const fieldSettings = {
+      detail: {
+        enabled: 'findings', // not used
+        required: 'always'
+      },
+      comment: {
+        enabled: 'always', // not used
+        required: 'always'
+      }
+    }
+
+    const allowAccept = true
+
+    const filePath = './WATCHER-test-files/WATCHER/ckl/no-Status-xml-element.ckl'
+
+    const review = await generateReviewObject(
+      filePath,
+      importOptions,
+      fieldSettings,
+      allowAccept
+    )
+
+    expect(review.checklists[0].reviews).to.be.empty
+  })
+  it('Testing no RuleID value in a ckl ', async () => {
+
+    const importOptions = {
+      autoStatus: 'submitted',
+      unreviewed: 'commented',
+      unreviewedCommented: 'informational',
+      emptyDetail: 'ignore',
+      emptyComment: 'ignore',
+      allowCustom: true
+    }
+
+    const fieldSettings = {
+      detail: {
+        enabled: 'findings', // not used
+        required: 'always'
+      },
+      comment: {
+        enabled: 'always', // not used
+        required: 'always'
+      }
+    }
+
+    const allowAccept = true
+
+    const filePath = './WATCHER-test-files/WATCHER/ckl/No-RuleId-Value.ckl'
+
+    const review = await generateReviewObject(
+      filePath,
+      importOptions,
+      fieldSettings,
+      allowAccept
+    )
+
+    expect(review.checklists[0].reviews).to.be.empty
+  })
 })
