+++ test/cklb-tests/CKLBReviewParserReviewObject.test.js
@@ -1,37 +1,11 @@
 import chai from 'chai';
-import { reviewsFromXccdf } from '../ReviewParser.js'; 
-import { XMLParser } from 'fast-xml-parser';
+import { reviewsFromCklb } from '../../ReviewParser.js'; 
 import fs from 'fs/promises';
-import he from 'he';
-const expect = chai.expect
-const valueProcessor = function (
-  tagName,
-  tagValue,
-  jPath,
-  hasAttributes,
-  isLeafNode
-) {
-  he.decode(tagValue)
-}
 
-const dataArray = [
-  {
-    scapBenchmarkId: 'CAN_Ubuntu_18-04_STIG',
-    benchmarkId: 'U_CAN_Ubuntu_18-04_STIG'
-  },
-  { scapBenchmarkId: 'Mozilla_Firefox_RHEL', benchmarkId: 'Mozilla_Firefox' },
-  {
-    scapBenchmarkId: 'Mozilla_Firefox_Windows',
-    benchmarkId: 'Mozilla_Firefox'
-  },
-  { scapBenchmarkId: 'MOZ_Firefox_Linux', benchmarkId: 'MOZ_Firefox_STIG' },
-  { scapBenchmarkId: 'MOZ_Firefox_Windows', benchmarkId: 'MOZ_Firefox_STIG' },
-  { scapBenchmarkId: 'Solaris_10_X86_STIG', benchmarkId: 'Solaris_10_X86' }
-]
 
-const scapBenchmarkMap = new Map(
-  dataArray.map(item => [item.scapBenchmarkId, item])
-)
+const expect = chai.expect;
+
+
 
 // Create a helper function to read the file and generate the review object
 async function generateReviewObject (
@@ -41,31 +15,33 @@ async function generateReviewObject (
   allowAccept
 ) {
   const data = await fs.readFile(filePath, 'utf8')
-  return reviewsFromXccdf({
+  return reviewsFromCklb({
     data,
     fieldSettings,
     allowAccept,
     importOptions,
-    valueProcessor,
-    scapBenchmarkMap,
-    XMLParser
   })
 }
 
-describe('Testing permutations of Import Options for a review objects parsed by xccdf Review Parser.', () => {
+
+describe('Import Options, allowAccept for a CKLb review object in non multi-stig', () => {
   it('DEFAULT SETTINGS: Primarily testing review "status = saved"', async () => {
     // Test: DEFAULT SETTINGS
-    // This test validates the behavior of the xccdf parser function under default settings.
+    // This test validates the behavior of the cklb parser function under default settings.
     // Primary Focus:
-    // - Ensuring that the 'autoStatus' option in 'importOptions' correctly sets the reviews to be empty
-  
+    // - Ensuring that the 'autoStatus' option in 'importOptions' correctly sets the review status to 'saved'.
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // These checks ensure that not only is the 'status' property set as expected, but also that
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-notReviewed-Commented-Detailed.cklb') to simulate a real-world scenario.
 
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
       unreviewedCommented: 'informational',
       emptyDetail: 'replace',
-      emptyComment: 'replace',
+      emptyComment: 'ignore',
       allowCustom: true
     }
 
@@ -83,23 +59,36 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-notReviewed-Commented-Detailed-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-notReviewed-Commented-Detailed.cklb'
+
     const review = await generateReviewObject(
       filePath,
       importOptions,
       fieldSettings,
       allowAccept
     )
+    
 
-    expect(review.checklists[0].reviews).to.be.empty
+    const expectedReview = {
+      ruleId: 'SV-257777r925317_rule',
+      status: 'saved',
+      result: 'informational',
+      comment: 'xyz',
+      detail: 'xyz'
+    }
+
+    expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
 
   it('autoStatus = null (keep exisiting), testing that all review statuses do not exist', async () => {
     // Test: autoStatus = null
-    // This test validates the behavior of the xccdf parser function under the 'autoStatus = null' setting.
+    // This test validates the behavior of the cklb parser function under the 'autoStatus = null' setting.
     // Primary Focus:
-    // - Ensuring that the 'autoStatus' option in 'importOptions' correctly sets the review to be empty
-  
+    // - Ensuring that the 'autoStatus' option in 'importOptions' correctly sets the review status to null (doesnt exist).
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'ruleId', 'result', 'comment', and 'detail'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-notReviewed-Commented-Detailed.cklb') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'null',
       unreviewed: 'commented',
@@ -123,7 +112,7 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-notReviewed-Commented-Detailed-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-notReviewed-Commented-Detailed.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -132,15 +121,28 @@ describe('Testing permutations of Import Options for a review objects parsed by 
       allowAccept
     )
 
-    expect(review.checklists[0].reviews).to.be.empty
+    const expectedReview = {
+      ruleId: 'SV-257777r925317_rule',
+      result: 'informational',
+      comment: 'xyz',
+      detail: 'xyz'
+    }
+
+    expect(review.checklists[0].reviews[0]).to.include(expectedReview)
+
+    expect(review.checklists[0].reviews[0].status).to.not.exist
   })
 
   it('autoStatus = submitted, testing if reviews are set to "submitted" if valid or "saved" if not valid. Determined by field settings and result', async () => {
     // Test: autoStatus = submitted
-    // This test validates the behavior of the ckl parser function under the autoStatus = submitted setting coupled with out field settings.
+    // This test validates the behavior of the cklb parser function under the autoStatus = submitted setting coupled with out field settings.
     // Primary Focus:
     // - Ensuring that the 'autoStatus' option in 'importOptions' correctly sets the review status to 'submitted' if valid or 'saved' if not valid.
-    
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including ruleId 'result', 'comment', and 'detail'.
+    // These checks ensure that not only is the 'status' property set as expected, but also that
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-notReviewed-Commented-Detailed.cklb') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -164,7 +166,7 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Asset_a-VPN_TRUNCATED-V2R5-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Asset_a-VPN_TRUNCATED-V2R5.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -207,8 +209,8 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     }
 
     // ensuring that each review has a status that matches the expected status
-    for (const checklist of review.checklists) {
-      for (const reviewItem of checklist.reviews) {
+    for (const checklists of review.checklists) {
+      for (const reviewItem of checklists.reviews) {
         const expectedStatus = expectedStatuses[reviewItem.ruleId]
         const expectedComment = expectedComments[reviewItem.ruleId]
         const expectedDetail = expectedDetails[reviewItem.ruleId]
@@ -219,12 +221,16 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     }
   })
 
-  it('autoStatus = submitted, testing that status is set to saved if does not meet field settings requirements', async () => {
+  it('autoStatus = submitted, testing that we will set status to saved if does not meet field settings requirements', async () => {
     // Test: autoStatus = submitted
-    // This test validates the behavior of the xccdf parser function under the autoStatus = submitted setting coupled with our field settings.
+    // This test validates the behavior of the cklb parser function under the autoStatus = submitted setting coupled with our field settings.
     // Primary Focus:
-    // - Ensuring that the 'autoStatus' option in 'importOptions' correctly sets the review to be empty
-
+    // - Ensuring that the 'autoStatus' option in 'importOptions' correctly sets the review status to 'saved' if not valid.
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail'.
+    // These checks ensure that not only is the 'status' property set as expected, but also that
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-notReviewed-Commented-Detailed.cklb') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -248,7 +254,7 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-notReviewed-Commented-Detailed-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-notReviewed-Commented-Detailed.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -256,19 +262,29 @@ describe('Testing permutations of Import Options for a review objects parsed by 
       fieldSettings,
       allowAccept
     )
-
-    expect(review.checklists[0].reviews).to.be.empty
+    const expectedReview = {
+      ruleId: 'SV-257777r925317_rule',
+      status: 'saved',
+      result: 'informational',
+      comment: 'xyz',
+      detail: 'xyz'
+    }
+    expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
 
   it('(autoStatus = accepted, allowAccept = true) for permissions to use accepted status', async () => {
     // Test: autoStatus = accepted, allowAccept = true
-    // This test validates the behavior of the xccdf parser function under the autoStatus = accepted, allowAccept = true settings.
+    // This test validates the behavior of the cklb parser function under the autoStatus = accepted, allowAccept = true settings.
     // Primary Focus:
     // - Ensuring that the 'autoStatus' option in 'importOptions' correctly sets the review status to 'submitted' if the user does not have permissions
     // or 'accepted' if the user does have permissions (permissions are determined by the 'allowAccept' option).
     // note: if not accepted or submitted, it will be saved because review did not meet the field settings requirements
     // note: in this test we will have permissions to accept reviews
-
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // These checks ensure that not only is the 'status' property set as expected, but also that
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Asset_a-VPN_TRUNCATED-V2R5.cklb'') to simulate a real-world scenario.
 
     const importOptions = {
       autoStatus: 'accepted',
@@ -292,7 +308,7 @@ describe('Testing permutations of Import Options for a review objects parsed by 
 
     const allowAccept = true
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Asset_a-VPN_TRUNCATED-V2R5-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Asset_a-VPN_TRUNCATED-V2R5.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -352,13 +368,17 @@ describe('Testing permutations of Import Options for a review objects parsed by 
 
   it('(autoStatus = accepted allowAccept = false) for permissions to use submitted status', async () => {
     // Test: autoStatus = accepted, allowAccept = false
-    // This test validates the behavior of the xccdf parser function under the autoStatus = accepted, allowAccept = false settings.
+    // This test validates the behavior of the cklb parser function under the autoStatus = accepted, allowAccept = false settings.
     // Primary Focus:
     // - Ensuring that the 'autoStatus' option in 'importOptions' correctly sets the review status to 'submitted' if the user does not have permissions
     // or 'accepted' if the user does have permissions (permissions are determined by the 'allowAccept' option).
     // note: if not accepted or submitted, it will be saved because review did not meet the field settings requirements
     // note: in this test we will NOT have permissions to accept reviews (we should see submitted)
-   
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // These checks ensure that not only is the 'status' property set as expected, but also that
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Asset_a-VPN_TRUNCATED-V2R5.cklb'') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'accepted',
       unreviewed: 'commented',
@@ -381,7 +401,7 @@ describe('Testing permutations of Import Options for a review objects parsed by 
 
     const allowAccept = false
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Asset_a-VPN_TRUNCATED-V2R5-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Asset_a-VPN_TRUNCATED-V2R5.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -438,13 +458,16 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     }
   })
 
-  it("'unreviewed = commented' testing that we ignore unreviewed reviews even if they have commment/detail  ", async () => {
+  it("'unreviewed = commented' testing that we only import unreviwred rules that contain a comment or detail   ", async () => {
     // Test: unreviewed = commented
-    // This test validates the behavior of the xccdf parser function under the unreviewed = commented settings.
+    // This test validates the behavior of the cklb parser function under the unreviewed = commented settings.
     // Primary Focus:
-    // - Ensuring that the 'unreviewed' option in 'importOptions' correctly ignores reviews that are unreviewed and have a comment or detail.
+    // - Ensuring that the 'unreviewed' option in 'importOptions' correctly imports only reviews with a non complience result that contain a comment and or detail.
     // note this test will have a review with a comment and detail
-
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-notReviewed-Commented-Detailed.cklb') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -468,7 +491,7 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-notReviewed-Commented-Detailed-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-notReviewed-Commented-Detailed.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -477,16 +500,33 @@ describe('Testing permutations of Import Options for a review objects parsed by 
       allowAccept
     )
 
-    expect(review.checklists[0].reviews).to.be.empty
+    const expectedReview = {
+      ruleId: 'SV-257777r925317_rule',
+      status: 'saved',
+      result: 'informational',
+      comment: 'xyz',
+      detail: 'xyz'
+    }
+    // check that review has a comment or detail and exists
+    expect(review.checklists[0].reviews[0]).to.exist
+    expect(
+      review.checklists[0].reviews[0].detail ||
+        review.checklists[0].reviews[0].comment
+    ).to.not.be.null
+    // secondary check that review matches expected review object
+    expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
 
-  it("'unreviewed = commented' testing that we do not import unreviwred rules that contain a comment or detail but giving a review without either", async () => {
+  it("'unreviewed = commented' testing that we only import unreviwred rules that contain a comment or detail but giving a review without either", async () => {
     // Test: unreviewed = commented
-    // This test validates the behavior of the xccdf parser function under the unreviewed = commented settings.
+    // This test validates the behavior of the cklb parser function under the unreviewed = commented settings.
     // Primary Focus:
     // - Ensuring that the 'unreviewed' option in 'importOptions' correctly imports only reviews with a non complience result that contain a comment and or detail.
     // note this test will not have any reviews because there will be no comment/detail
-    
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-notReviewed-Commented-Detailed.cklb') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -510,7 +550,7 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-notReviewed-Empty-CommentDetail-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-notReviewed-Empty-CommentDetail.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -523,13 +563,16 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     expect(review.checklists[0].reviews).to.be.empty
   })
 
-  it("'unreviewed = always', testing that unreviewed always are import and result is 'notchecked'  ", async () => {
+  it("'unreviewed = always', testing that unreviewed always are import  ", async () => {
     // Test: unreviewed = always
-    // This test validates the behavior of the xccdf parser function under the unreviewed = always settings.
+    // This test validates the behavior of the cklb parser function under the unreviewed = always settings.
     // Primary Focus:
     // - Ensuring that the 'unreviewed' option in 'importOptions' correctly imports all reviews without a compliance result and if they have a comment
-    // or detail they will be labeled as notchecked.
-
+    // or detail they will be labeled as informational.
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-notReviewed-Commented-Detailed.cklb') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'always',
@@ -553,7 +596,7 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-notReviewed-Commented-Detailed-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-notReviewed-Commented-Detailed.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -563,9 +606,9 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     )
 
     const expectedReview = {
-      ruleId: 'SV-2',
+      ruleId: 'SV-257777r925317_rule',
       status: 'saved',
-      result: 'notchecked',
+      result: 'informational',
       comment: 'xyz',
       detail: 'xyz'
     }
@@ -573,13 +616,16 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
 
-  it("'unreviewed = always', testing that unreviewed always are imported ", async () => {
+  it("'unreviewed = always', testing that unreviewed always are import  ", async () => {
     // Test: unreviewed = always
-    // This test validates the behavior of the xccdf parser function under the unreviewed = always settings.
+    // This test validates the behavior of the cklb parser function under the unreviewed = always settings.
     // Primary Focus:
     // - Ensuring that the 'unreviewed' option in 'importOptions' correctly imports all reviews without a compliance result
     // and without comment or detail to be labled as notchecked.
-
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // These checks ensure that not only is the 'status' property set as expected, but also that
+    // The test utilizes a sample CKLb file ('Single-Vuln-notReviewed-Empty-CommentDetail.cklb') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'always',
@@ -603,7 +649,7 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-notReviewed-Empty-CommentDetail-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-notReviewed-Empty-CommentDetail.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -611,8 +657,9 @@ describe('Testing permutations of Import Options for a review objects parsed by 
       fieldSettings,
       allowAccept
     )
+   
     const expectedReview = {
-      ruleId: 'SV-2',
+      ruleId: 'SV-257777r925317_rule',
       status: 'saved',
       result: 'notchecked',
       comment: null,
@@ -624,10 +671,13 @@ describe('Testing permutations of Import Options for a review objects parsed by 
 
   it(" 'unreviewed = never' testing to never import an unreviewed item ", async () => {
     // Test: unreviewed = never
-    // This test validates the behavior of the xccdf parser function under the unreviewed = never settings.
+    // This test validates the behavior of the cklb parser function under the unreviewed = never settings.
     // Primary Focus:
     // - Ensuring that the 'unreviewed' option in 'importOptions' ignores reviews without a compliance result (Nf/na/o)
-    
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-notReviewed-Empty-CommentDetail.cklb') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'never',
@@ -651,7 +701,7 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-notReviewed-Empty-CommentDetail-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-notReviewed-Empty-CommentDetail.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -663,12 +713,16 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     expect(review.checklists[0].reviews).to.be.empty
   })
 
-  it("'unreviewedComment = informational' testing that an unreviewed item with a comment is an empty review ", async () => {
+  it("'unreviewedComment = informational' testing that an unreviewed item with a comment has a result of informational", async () => {
     // Test: unreviewedComment = informational
-    // This test validates the behavior of the xccdf parser function under the unreviewedComment = informational settings.
+    // This test validates the behavior of the cklb parser function under the unreviewedComment = informational settings.
     // Primary Focus:
-    // - Ensuring that the 'unreviewed' option in 'importOptions' correctly does not import the review.
-    
+    // - Ensuring that the 'unreviewed' option in 'importOptions' correctly labels a review without a compliance result..
+    // and with a comment or detail as informational.
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-notReviewed-Commented-Detailed.cklb') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -691,7 +745,7 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-notReviewed-Commented-Detailed-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-notReviewed-Commented-Detailed.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -700,15 +754,26 @@ describe('Testing permutations of Import Options for a review objects parsed by 
       allowAccept
     )
 
-    expect(review.checklists[0].reviews).to.be.empty
+    const expectedReview = {
+      ruleId: 'SV-257777r925317_rule',
+      status: 'saved',
+      result: 'informational',
+      comment: 'xyz',
+      detail: 'xyz'
+    }
+
+    expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
-
-  it(" 'unreviewedComment = notchecked'. testing that an unreviewed with a comment returns no reviews", async () => {
+  it(" 'unreviewedComment = notchecked'. testing that an unreviewed with a comment has a result of notchecked", async () => {
     // Test: unreviewedComment = notchecked
-    // This test validates the behavior of the xccdf parser function under the unreviewedComment = notchecked settings.
+    // This test validates the behavior of the cklb parser function under the unreviewedComment = notchecked settings.
     // Primary Focus:
-    // - Ensuring that the 'unreviewed' option in 'importOptions' correctly ignores a review without a compliance result
-
+    // - Ensuring that the 'unreviewed' option in 'importOptions' correctly labels a review without a compliance result..
+    // and with a comment or detail as notchecked.
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-notReviewed-Commented-Detailed.cklb') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -731,7 +796,7 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-notReviewed-Commented-Detailed-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-notReviewed-Commented-Detailed.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -739,15 +804,26 @@ describe('Testing permutations of Import Options for a review objects parsed by 
       fieldSettings,
       allowAccept
     )
+    const expectedReview = {
+      ruleId: 'SV-257777r925317_rule',
+      status: 'saved',
+      result: 'notchecked',
+      comment: 'xyz',
+      detail: 'xyz'
+    }
 
-    expect(review.checklists[0].reviews).to.be.empty
+    expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
 
   it("'emptyDetail = replace' testing that if an item has an empty detail we will replace it with a static message", async () => {
     // Test: emptyDetail = replace
-    // This test validates the behavior of the xccdf parser function under the emptyDetail = replace settings.
+    // This test validates the behavior of the cklb parser function under the emptyDetail = replace settings.
     // Primary Focus:
-    // - Ensuring that the 'emptyDetail' option in 'importOptions' correctly replaces an empty detail with a static message.
+    // - Ensuring that the 'emptyDetail' option in 'importOptions' correctly replaces an empty detail with a static message. "There is no detail provided for the assessment"
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-Pass-With-Comment.cklb') to simulate a real-world scenario.
 
     const importOptions = {
       autoStatus: 'saved',
@@ -771,7 +847,7 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-Pass-With-Comment-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-Pass-With-Comment.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -780,21 +856,24 @@ describe('Testing permutations of Import Options for a review objects parsed by 
       allowAccept
     )
     const expectedReview = {
-      ruleId: 'SV-2',
+      ruleId: 'SV-207191r803417_rule',
       status: 'saved',
       result: 'pass',
       comment: 'xyz',
-      detail:
-        'Result was reported by product "MyTestSystem" version 1 at 2023-11-13T16:41:49.000Z using check content "MyCheckContent"'
+      detail: 'There is no detail provided for the assessment'
     }
+
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
-
-  it("'emptyDetail = ignore' testing that if there is no detail provided it will retain exisiting detail, if no exisitng then we will set to null", async () => {
+  it("'emptyDetail = ignore' testing that if there is no detail provided it will retaing exisiting, if no exisitng then we will set to null", async () => {
     // Test: emptyDetail = ignore
-    // This test validates the behavior of the xccdf parser function under the emptyDetail = ignore settings.
+    // This test validates the behavior of the cklb parser function under the emptyDetail = ignore settings.
     // Primary Focus:
     // - Ensuring that the 'emptyDetail' option in 'importOptions' correctly replaces an empty detail a null value
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-Pass-With-Comment.cklb') to simulate a real-world scenario.
 
     const importOptions = {
       autoStatus: 'saved',
@@ -818,7 +897,7 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-Pass-With-Comment-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-Pass-With-Comment.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -828,7 +907,7 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     )
 
     const expectedReview = {
-      ruleId: 'SV-2',
+      ruleId: 'SV-207191r803417_rule',
       status: 'saved',
       result: 'pass',
       comment: 'xyz',
@@ -839,10 +918,13 @@ describe('Testing permutations of Import Options for a review objects parsed by 
   })
   it("'emptyDetail = import', testing empty detail will clear existing text (setting it to empty string)", async () => {
     // Test: emptyDetail = import
-    // This test validates the behavior of the xccdf parser function under the emptyDetail = import settings.
+    // This test validates the behavior of the cklb parser function under the emptyDetail = import settings.
     // Primary Focus:
     // - Ensuring that the 'emptyDetail' option in 'importOptions' correctly replaces an empty detail an empty string if no detail is provided
-   
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-Pass-With-Comment.cklb') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -865,7 +947,7 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-Pass-With-Comment-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-Pass-With-Comment.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -875,7 +957,7 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     )
 
     const expectedReview = {
-      ruleId: 'SV-2',
+      ruleId: 'SV-207191r803417_rule',
       status: 'saved',
       result: 'pass',
       comment: 'xyz',
@@ -887,10 +969,13 @@ describe('Testing permutations of Import Options for a review objects parsed by 
 
   it(" 'emptyDetail = import' testing that a review with a detail provided will be applied", async () => {
     // Test: emptyDetail = import
-    // This test validates the behavior of the xccdf parser function under the emptyDetail = import settings.
+    // This test validates the behavior of the cklb parser function under the emptyDetail = import settings.
     // Primary Focus:
     // - Ensuring that the 'emptyDetail' option in 'importOptions' correctly uses the exisisitng detail if one is provided
-  
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-Pass-With-Detail.cklb') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -913,7 +998,7 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-Pass-With-Detail-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-Pass-With-Detail.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -923,7 +1008,7 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     )
 
     const expectedReview = {
-      ruleId: 'SV-2',
+      ruleId: 'SV-207191r803417_rule',
       status: 'saved',
       result: 'pass',
       comment: null,
@@ -932,12 +1017,15 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
 
-  it("'emptyComment = replace' testing that if an item has an empty comment we will replace it with a dynamically created message", async () => {
+  it("'emptyComment = replace' testing that if an item has an empty comment we will replace it with a static message", async () => {
     // Test: emptyComment = replace
-    // This test validates the behavior of the xxcdf parser function under the emptyComment = replace settings.
+    // This test validates the behavior of the cklb parser function under the emptyComment = replace settings.
     // Primary Focus:
-    // - Ensuring that the 'emptyComment' option in 'importOptions' correctly replaces an empty comment with a dynamically created message based off the data.
-   
+    // - Ensuring that the 'emptyComment' option in 'importOptions' correctly replaces an empty comment with a static message. "There is no comment provided for the assessment"
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-Pass-With-Detail.cklb') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -960,7 +1048,7 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-Pass-With-Detail-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-Pass-With-Detail.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -970,11 +1058,10 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     )
 
     const expectedReview = {
-      ruleId: 'SV-2',
+      ruleId: 'SV-207191r803417_rule',
       status: 'saved',
       result: 'pass',
-      comment:
-        'Result was reported by product "MyTestSystem" version 1 at 2023-11-13T16:41:49.000Z using check content "MyCheckContent"',
+      comment: 'There is no comment provided for the assessment',
       detail: 'xyz'
     }
 
@@ -983,10 +1070,13 @@ describe('Testing permutations of Import Options for a review objects parsed by 
 
   it("'emptyComment = ignore' testing that we will use exisitng text, if none use null", async () => {
     // Test: emptyComment = ignore
-    // This test validates the behavior of the xccdf parser function under the emptyComment = ignore settings.
+    // This test validates the behavior of the cklb parser function under the emptyComment = ignore settings.
     // Primary Focus:
     // - Ensuring that the 'emptyComment' option in 'importOptions' correctly replaces an empty comment a null value
-  
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-Pass-With-Comment.cklb') to simulate a real-world scenario.
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -1009,7 +1099,7 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-Pass-With-Detail-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-Pass-With-Detail.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -1018,7 +1108,7 @@ describe('Testing permutations of Import Options for a review objects parsed by 
       allowAccept
     )
     const expectedReview = {
-      ruleId: 'SV-2',
+      ruleId: 'SV-207191r803417_rule',
       status: 'saved',
       result: 'pass',
       comment: null,
@@ -1029,10 +1119,13 @@ describe('Testing permutations of Import Options for a review objects parsed by 
 
   it("'emptyComment = import', will clear eixsitng with an empty string if no comment given ", async () => {
     // Test: emptyComment = import
-    // This test validates the behavior of the xccdf parser function under the emptyComment = import settings.
+    // This test validates the behavior of the cklb parser function under the emptyComment = import settings.
     // Primary Focus:
     // - Ensuring that the 'emptyComment' option in 'importOptions' correctly replaces an empty comment an empty string: ""
-    
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-Pass-With-Detail.cklb')
 
     const importOptions = {
       autoStatus: 'saved',
@@ -1056,7 +1149,7 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-Pass-With-Detail-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-Pass-With-Detail.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -1066,7 +1159,7 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     )
 
     const expectedReview = {
-      ruleId: 'SV-2',
+      ruleId: 'SV-207191r803417_rule',
       status: 'saved',
       result: 'pass',
       comment: '',
@@ -1075,12 +1168,15 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
 
-  it("'emptyComment = import', testing a review with a comment provided in the ckl to make sure we get it back in the review ", async () => {
+  it("'emptyComment = import', testing a review with a comment provided in the cklb to make sure we get it back in the review ", async () => {
     // Test: emptyComment = import
-    // This test validates the behavior of the xccdf parser function under the emptyComment = import settings.
+    // This test validates the behavior of the cklb parser function under the emptyComment = import settings.
     // Primary Focus:
     // - Ensuring that the 'emptyComment' option in 'importOptions' correctly uses the exisisitng comment if one is provided
- 
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-Pass-With-Detail.cklb')
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -1103,7 +1199,7 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-Pass-With-Comment-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-Pass-With-Comment.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -1113,7 +1209,7 @@ describe('Testing permutations of Import Options for a review objects parsed by 
     )
 
     const expectedReview = {
-      ruleId: 'SV-2',
+      ruleId: 'SV-207191r803417_rule',
       status: 'saved',
       result: 'pass',
       comment: 'xyz',
@@ -1123,13 +1219,16 @@ describe('Testing permutations of Import Options for a review objects parsed by 
   })
 })
 
-describe('fieldSettings testing for a review object in non multi-stig xccdf', () => {
+ describe('fieldSettings testing for a CKLb review object in non multi-stig', () => {
   it("DEFAULT FIELD SETTINGS with allowAccept=true and a passing review, testing that it has a detail and is 'submitted'", async () => {
     // Test: autostatus = submitted, default field settings.
-    // This test validates the behavior of the xccdf parser function under the default field settings.
+    // This test validates the behavior of the cklb parser function under the default field settings.
     // Primary Focus:
     // - Ensuring that the 'autoStatus' option in 'importOptions' correctly ensures that a detail is required for a review to be submitted.
-   
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-Pass-With-Detail.cklb')
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1153,7 +1252,7 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-Pass-With-Detail-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-Pass-With-Detail.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -1163,7 +1262,7 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
     )
 
     const expectedReview = {
-      ruleId: 'SV-2',
+      ruleId: 'SV-207191r803417_rule',
       status: 'submitted',
       result: 'pass',
       comment: '',
@@ -1173,14 +1272,16 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
     // expected status is submitted for the rule that has a detail
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
-
   it('DEFAULT FIELD SETTINGS  with allowAccept=true and a failing review with no detail.', async () => {
     //autostatus = submitted, testing default field settings with allowAccept=true
-    // This test validates the behavior of the xccdf parser function under the default field setting with a fail and a no detail.
+    // This test validates the behavior of the cklb parser function under the default field setting with a fail and a no detail.
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a detail is always required for submission for a review.
     // Test that with a failing review and no detail exisiting and it will be set to 'saved'
-   
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-fail-With-Detail.cklb')
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1204,7 +1305,7 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-fail-Empty-CommentDetail-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-fail-Empty-CommentDetail.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -1214,7 +1315,7 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
     )
 
     const expectedReview = {
-      ruleId: 'SV-2',
+      ruleId: 'SV-207191r803417_rule',
       status: 'saved',
       result: 'fail',
       comment: null,
@@ -1223,14 +1324,16 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
 
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
-
   it('fieldSettings.detail.required = findings with allowAccept=true with a failing review containing a detail', async () => {
     // Test: autostatus = submitted, default field settings.
-    // This test validates the behavior of the xccdf parser function under the  fieldSettings.detail.required = findings field settings.
+    // This test validates the behavior of the cklb parser function under the  fieldSettings.detail.required = findings field settings.
     // Primary Focus:
     // - Ensuring that the 'autoStatus' option in 'importOptions' correctly ensures that a detail is required for a review  that has findings to be submitted.
     // Test that a failing review with a detail will be submitted
-   
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-Pass-With-Detail.cklb')
 
     const importOptions = {
       autoStatus: 'submitted',
@@ -1255,7 +1358,7 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-fail-With-Detail-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-fail-With-Detail.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -1265,7 +1368,7 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
     )
 
     const expectedReview = {
-      ruleId: 'SV-2',
+      ruleId: 'SV-207191r803417_rule',
       status: 'submitted',
       result: 'fail',
       comment: '',
@@ -1274,14 +1377,16 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
 
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
-
   it("'fieldSettings.detail.required = findings' with allowAccept=true with a fail and no detail or comment", async () => {
     // Test: autostatus = submitted, testing 'fieldSettings.detail.required = findings' allowAccept=true
-    // This test validates the behavior of the xccdf parser function under the  fieldSettings.detail.required = findings field setting with a fail and no detail or comment.
+    // This test validates the behavior of the cklb parser function under the  fieldSettings.detail.required = findings field setting with a fail and no detail or comment.
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a detail is required for a review that has findings to be submitted if not we will save.
     // Test that no detail exisitng and it will be set to 'saved'
-   
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-fail-Empty-CommentDetail.cklb')
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1305,7 +1410,7 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-fail-Empty-CommentDetail-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-fail-Empty-CommentDetail.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -1314,23 +1419,28 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
       allowAccept
     )
     const expectedReview = {
-      ruleId: 'SV-2',
+      ruleId: 'SV-207191r803417_rule',
       status: 'saved',
       result: 'fail',
       comment: null,
       detail: null
     }
 
+    // console.log(JSON.stringify(review, null, 2))
+
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
 
   it("'fieldSettings.detail.required = optional' with allowAccept=true with a fail and no detail or comment, testing that it does not have a detail and is submitted ", async () => {
     // TEST: autostatus = submitted, testing 'fieldSettings.detail.required = optional' with allowAccept=true
-    // This test validates the behavior of the xccdf parser function under the  fieldSettings.detail.required = optional field setting with a fail and no detail or comment.
+    // This test validates the behavior of the cklb parser function under the  fieldSettings.detail.required = optional field setting with a fail and no detail or comment.
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a detail is optional for submission for a review that has findings.
     // Test that no detail exisitng and it will be set to 'submitted'
-   
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-fail-Empty-CommentDetail.cklb')
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1354,7 +1464,7 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-fail-Empty-CommentDetail-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-fail-Empty-CommentDetail.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -1364,7 +1474,7 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
     )
 
     const expectedReview = {
-      ruleId: 'SV-2',
+      ruleId: 'SV-207191r803417_rule',
       status: 'submitted',
       result: 'fail',
       comment: null,
@@ -1373,14 +1483,16 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
 
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
-
   it("fieldSettings.detail.required = optional' with allowAccept=true with a fail and detail testing it has a detail and is submitted", async () => {
     // autostatus = submitted, 'fieldSettings.detail.required = optional' with allowAccept=true
-    // This test validates the behavior of the xccdf parser function under the  fieldSettings.detail.required = optional field setting with a fail and a detail.
+    // This test validates the behavior of the cklb parser function under the  fieldSettings.detail.required = optional field setting with a fail and a detail.
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a detail is optional for submission for a review that has findings.
     // Test that with detail exisitng and it will be set to 'submitted'
-   
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-fail-With-Detail.cklb')
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1404,7 +1516,7 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-fail-With-Detail-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-fail-With-Detail.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -1414,7 +1526,7 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
     )
 
     const expectedReview = {
-      ruleId: 'SV-2',
+      ruleId: 'SV-207191r803417_rule',
       status: 'submitted',
       result: 'fail',
       comment: null,
@@ -1426,11 +1538,14 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
 
   it('DEFAULT FIELD SETTINGS with allowAccept=true and a passing review testing that it has a comment and is submitted', async () => {
     //autostatus = submitted, testing default field settings with allowAccept=true
-    // This test validates the behavior of the xccdf parser function under the default field setting with a pass and a comment.
+    // This test validates the behavior of the cklb parser function under the default field setting with a pass and a comment.
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a comment is always required for submission for a review.
     // Test that witha passing review and a comment exisiting and it will be set to 'submitted'
-   
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-fail-With-Detail.cklb')
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1454,7 +1569,7 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-Pass-With-Comment-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-Pass-With-Comment.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -1464,7 +1579,7 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
     )
 
     const expectedReview = {
-      ruleId: 'SV-2',
+      ruleId: 'SV-207191r803417_rule',
       status: 'submitted',
       result: 'pass',
       comment: 'xyz',
@@ -1473,14 +1588,16 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
 
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
-
   it('DEFAULT FIELD SETTINGS  with allowAccept=true and a failing review with no comment.', async () => {
     //autostatus = submitted, testing default field settings with allowAccept=true
-    // This test validates the behavior of the xccdf parser function under the default field setting with a fail and a no comment.
+    // This test validates the behavior of the cklb parser function under the default field setting with a fail and a no comment.
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a comment is always required for submission for a review.
     // Test that with a failing review and no comment exisiting and it will be set to 'saved'
-    
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-fail-With-Detail.cklb')
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1504,7 +1621,7 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-fail-Empty-CommentDetail-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-fail-Empty-CommentDetail.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -1514,7 +1631,7 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
     )
 
     const expectedReview = {
-      ruleId: 'SV-2',
+      ruleId: 'SV-207191r803417_rule',
       status: 'saved',
       result: 'fail',
       comment: null,
@@ -1525,11 +1642,14 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
   })
   it('fieldSettings.comment.required = findings with allowAccept=true with a fail and comment testing that it has a comment and is submitted', async () => {
     // TEST: autostatus = submitted, fieldSettings.comment.required = findings with allowAccept=true
-    // This test validates the behavior of the xccdf parser function under the fieldSettings.comment.required = findings field setting with a fail and a comment.
+    // This test validates the behavior of the cklb parser function under the fieldSettings.comment.required = findings field setting with a fail and a comment.
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a comment is always required for submission for a review that contains a finding.
     // Test that with a failed review and a comment exisitng and it will be set to 'submitted'
-    
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-fail-with-Comment.cklb')
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1553,7 +1673,7 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-fail-with-Comment-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-fail-with-Comment.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -1563,7 +1683,7 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
     )
 
     const expectedReview = {
-      ruleId: 'SV-2',
+      ruleId: 'SV-207191r803417_rule',
       status: 'submitted',
       result: 'fail',
       comment: 'xyz',
@@ -1574,11 +1694,14 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
   })
   it("'fieldSettings.comment.required = findings' with allowAccept=true with a fail and no detail or comment, testing that it does not have a comment and is 'saved' ", async () => {
     // TEST: autostatus = submitted, testing 'fieldSettings.comment.required = findings' with allowAccept=true
-    // This test validates the behavior of the xccdf parser function under the fieldSettings.comment.required = findings field setting with a fail and no comment.
+    // This test validates the behavior of the cklb parser function under the fieldSettings.comment.required = findings field setting with a fail and no comment.
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a comment is always required for submission for a review that contains a finding.
     // Test that with a failed review and no comment will be set to 'saved'
-    
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-fail-Empty-CommentDetail.cklb')
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1602,7 +1725,7 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-fail-Empty-CommentDetail-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-fail-Empty-CommentDetail.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -1612,7 +1735,7 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
     )
 
     const expectedReview = {
-      ruleId: 'SV-2',
+      ruleId: 'SV-207191r803417_rule',
       status: 'saved',
       result: 'fail',
       comment: null,
@@ -1624,11 +1747,14 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
 
   it("'fieldSettings.comment.required = optional' with allowAccept=true with a fail and no detail or comment. testing that it doesnt have a comment and is submmited", async () => {
     // TEST: autostatus = submitted, testing 'fieldSettings.comment.required = optional' with allowAccept=true
-    // This test validates the behavior of the xccdf parser function under the fieldSettings.comment.required = optional field setting with a fail and no comment or detail .
+    // This test validates the behavior of the cklb parser function under the fieldSettings.comment.required = optional field setting with a fail and no comment or detail .
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a comment is optionally required for submission for a review that contains a finding.
     // Test that with a failed review and no comment will be set to 'submitted'
- 
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-fail-Empty-CommentDetail.cklb')
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1652,7 +1778,7 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-fail-Empty-CommentDetail-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-fail-Empty-CommentDetail.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -1662,7 +1788,7 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
     )
 
     const expectedReview = {
-      ruleId: 'SV-2',
+      ruleId: 'SV-207191r803417_rule',
       status: 'submitted',
       result: 'fail',
       comment: null,
@@ -1673,11 +1799,14 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
   })
   it("fieldSettings.comment.required = optional' with allowAccept=true with a fail and comment, testing thhat it has a comment and is submitted", async () => {
     // TEST: autostatus = submitted, 'fieldSettings.comment.required = optional' with allowAccept=true
-    // This test validates the behavior of the xccdf parser function under the fieldSettings.comment.required = optional field setting with a fail and a comment.
+    // This test validates the behavior of the cklb parser function under the fieldSettings.comment.required = optional field setting with a fail and a comment.
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a comment is optionally required for submission for a review that contains a finding.
     // Test that with a failed review and a comment will be set to 'submitted'
-
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-fail-with-Comment.cklb')
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1701,7 +1830,7 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-fail-with-Comment-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-fail-with-Comment.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -1711,7 +1840,7 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
     )
 
     const expectedReview = {
-      ruleId: 'SV-2',
+      ruleId: 'SV-207191r803417_rule',
       status: 'submitted',
       result: 'fail',
       comment: 'xyz',
@@ -1720,16 +1849,19 @@ describe('fieldSettings testing for a review object in non multi-stig xccdf', ()
 
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
-})
+ })
 
-describe('Tests where fieldSettings and importOptions overlap xccdf. ', () => {
+describe(' CKLb Tests where fieldSettings and importOptions overlap. ', () => {
   it("Testing where emptyDetail: 'ignore', emptyComment: 'ignore', aswell as requiring a comment and detail  ", async () => {
     // TEST: emptyDetail: 'ignore', emptyComment: 'ignore', fieldSettings.detail.required = always, fieldSettings.comment.required = always
-    // This test validates the behavior of the xccdf parser function under above settings with a non compliance resilt and no comment or detail.
+    // This test validates the behavior of the cklb parser function under above settings with a non compliance resilt and no comment or detail.
     // Primary Focus:
     // - ensuring that we will have no reviews because a brand new review will be created with "null" comment or detail with are both required.
     // Test that with a failed review and a comment will be set to 'submitted'
-   
+    // Secondary Focus:
+    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
+    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
+    // The test utilizes a sample CKLb file ('Single-Vuln-fail-with-Comment.cklb')
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1753,7 +1885,7 @@ describe('Tests where fieldSettings and importOptions overlap xccdf. ', () => {
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-notReviewed-Empty-CommentDetail-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-notReviewed-Empty-CommentDetail.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -1766,12 +1898,14 @@ describe('Tests where fieldSettings and importOptions overlap xccdf. ', () => {
   })
 })
 
-describe('MISC. xccdf ', () => {
-  it('review with result engine data', async () => {
+describe('MISC CKLb. ', () => {
+  it('Testing that long comment.detail is truncated ', async () => {
+    // NOTE: the input comment and detail are '32768' characters long
+
     const importOptions = {
-      autoStatus: 'saved',
-      unreviewed: 'never',
-      unreviewedCommented: 'notchecked',
+      autoStatus: 'submitted',
+      unreviewed: 'commented',
+      unreviewedCommented: 'informational',
       emptyDetail: 'ignore',
       emptyComment: 'ignore',
       allowCustom: true
@@ -1779,19 +1913,19 @@ describe('MISC. xccdf ', () => {
 
     const fieldSettings = {
       detail: {
-        enabled: 'always',
+        enabled: 'findings', // not used
         required: 'always'
       },
       comment: {
-        enabled: 'findings',
-        required: 'findings'
+        enabled: 'always', // not used
+        required: 'always'
       }
     }
 
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-Pass-With-Comment-xccdf.xml'
+      './WATCHER-test-files/WATCHER/cklb/Single-Vuln-Long-CommentDetail.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -1800,73 +1934,10 @@ describe('MISC. xccdf ', () => {
       allowAccept
     )
 
-    //  console.log(JSON.stringify(review, null, 2))
+    const maxLength = 32767
 
-    const expectedResultEngineReview = {
-      ruleId: 'SV-2',
-      result: 'pass',
-      resultEngine: {
-        time: '2023-11-13T16:41:49.000Z',
-        type: 'scap',
-        version: "1",
-        product: 'MyTestSystem',
-        checkContent: {
-          location: 'MyCheckContent',
-          component: 'fso_comp_MyStig:135'
-        }
-      },
-      detail: null,
-      comment: 'xyz',
-      status: 'saved'
-    }
-
-    expect(review.checklists[0].reviews[0]).to.deep.equal(
-      expectedResultEngineReview
-    )
+    expect(review.checklists[0].reviews[0].detail).to.have.lengthOf(maxLength)
+    expect(review.checklists[0].reviews[0].comment).to.have.lengthOf(maxLength)
   })
-  it('review with result override', async () => {
-    const importOptions = {
-      autoStatus: 'saved',
-      unreviewed: 'never',
-      unreviewedCommented: 'notchecked',
-      emptyDetail: 'ignore',
-      emptyComment: 'ignore',
-      allowCustom: true
-    }
-
-    const fieldSettings = {
-      detail: {
-        enabled: 'always',
-        required: 'always'
-      },
-      comment: {
-        enabled: 'findings',
-        required: 'findings'
-      }
-    }
-
-    const allowAccept = true
-
-    const filePath =
-      './WATCHER-test-files/WATCHER/xccdf/ReviewOverrides-xccdf.xml'
-
-    const review = await generateReviewObject(
-      filePath,
-      importOptions,
-      fieldSettings,
-      allowAccept
-    )
-
-
-    const expectedOveride = {
-      authority: 'Authority1',
-      oldResult: 'pass',
-      newResult: 'fail',
-      remark: 'Some remark'
-    }
-
-    expect(review.checklists[0].reviews[0].resultEngine.overrides[0]).to.deep.equal(
-      expectedOveride
-    )
-  })
+  
 })
