+++ test/cklb-tests/CKLBStatisticsObject.test.js
@@ -1,9 +1,8 @@
 import chai from 'chai';
-import { reviewsFromXccdf } from '../ReviewParser.js'; 
+import { reviewsFromCklb } from '../../ReviewParser.js';  
 import { XMLParser } from 'fast-xml-parser';
 import fs from 'fs/promises';
 import he from 'he';
-
 const expect = chai.expect
 const valueProcessor = function (
   tagName,
@@ -15,25 +14,6 @@ const valueProcessor = function (
   he.decode(tagValue)
 }
 
-const dataArray = [
-  {
-    scapBenchmarkId: 'CAN_Ubuntu_18-04_STIG',
-    benchmarkId: 'U_CAN_Ubuntu_18-04_STIG'
-  },
-  { scapBenchmarkId: 'Mozilla_Firefox_RHEL', benchmarkId: 'Mozilla_Firefox' },
-  {
-    scapBenchmarkId: 'Mozilla_Firefox_Windows',
-    benchmarkId: 'Mozilla_Firefox'
-  },
-  { scapBenchmarkId: 'MOZ_Firefox_Linux', benchmarkId: 'MOZ_Firefox_STIG' },
-  { scapBenchmarkId: 'MOZ_Firefox_Windows', benchmarkId: 'MOZ_Firefox_STIG' },
-  { scapBenchmarkId: 'Solaris_10_X86_STIG', benchmarkId: 'Solaris_10_X86' }
-]
-
-const scapBenchmarkMap = new Map(
-  dataArray.map(item => [item.scapBenchmarkId, item])
-)
-
 // Create a helper function to read the file and generate the review object
 async function generateReviewObject (
   filePath,
@@ -42,20 +22,21 @@ async function generateReviewObject (
   allowAccept
 ) {
   const data = await fs.readFile(filePath, 'utf8')
-  return reviewsFromXccdf({
+  return reviewsFromCklb({
     data,
+    importOptions,
     fieldSettings,
     allowAccept,
-    importOptions,
     valueProcessor,
-    scapBenchmarkMap,
     XMLParser
   })
 }
 
 
-describe('Testing that the xccdf Review Parser will return the correct figures in the Statistics object', () => {
+describe('Testing that the CKLb Review Parser will return the correct figures in the Statistics object', () => {
   it('unreviewed: commented, unreviewedCommented: informational, has comments/detail', async () => {
+    // will import commented unreviewed findings as informational
+    // expecting to see 1 informational finding
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -78,7 +59,7 @@ describe('Testing that the xccdf Review Parser will return the correct figures i
 
     const allowAccept = true
 
-    const filePath = './WATCHER-test-files/WATCHER/xccdf/GoodStatistics-xccdf.xml'
+    const filePath = './WATCHER-test-files/WATCHER/cklb/GoodStatistics.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -93,7 +74,7 @@ describe('Testing that the xccdf Review Parser will return the correct figures i
       notapplicable: 2,
       notchecked: 0,
       notselected: 0,
-      informational: 0,
+      informational: 1,
       error: 0,
       fixed: 0,
       unknown: 0
@@ -103,6 +84,7 @@ describe('Testing that the xccdf Review Parser will return the correct figures i
   })
 
   it('unreviewed: commented, unreviewedCommented: notchecked, has comments/detail', async () => {
+     // will import commented unreviewed findings as notchecked
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -125,7 +107,7 @@ describe('Testing that the xccdf Review Parser will return the correct figures i
 
     const allowAccept = true
 
-    const filePath = './WATCHER-test-files/WATCHER/xccdf/GoodStatistics-xccdf.xml'
+    const filePath = './WATCHER-test-files/WATCHER/cklb/GoodStatistics.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -138,7 +120,7 @@ describe('Testing that the xccdf Review Parser will return the correct figures i
       pass: 2,
       fail: 2,
       notapplicable: 2,
-      notchecked: 0,
+      notchecked: 1,
       notselected: 0,
       informational: 0,
       error: 0,
@@ -150,6 +132,7 @@ describe('Testing that the xccdf Review Parser will return the correct figures i
   })
 
   it('unreviewed: always, unreviewedCommented: informational, has comments/detail', async () => {
+    // will always import unreviewed findings and unreviewed with a commment/detail is informational
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'always',
@@ -172,7 +155,7 @@ describe('Testing that the xccdf Review Parser will return the correct figures i
 
     const allowAccept = true
 
-    const filePath = './WATCHER-test-files/WATCHER/xccdf/GoodStatistics-xccdf.xml'
+    const filePath = './WATCHER-test-files/WATCHER/cklb/GoodStatistics.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -184,9 +167,9 @@ describe('Testing that the xccdf Review Parser will return the correct figures i
       pass: 2,
       fail: 2,
       notapplicable: 2,
-      notchecked: 2,
+      notchecked: 1,
       notselected: 0,
-      informational: 0,
+      informational: 1,
       error: 0,
       fixed: 0,
       unknown: 0
@@ -196,6 +179,7 @@ describe('Testing that the xccdf Review Parser will return the correct figures i
   })
 
   it('unreviewed: always, unreviewedCommented: notchecked, has comments/detail', async () => {
+      // will always import unreviewed findings and unreviewed with a commment/detail is notchecked
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'always',
@@ -218,7 +202,7 @@ describe('Testing that the xccdf Review Parser will return the correct figures i
 
     const allowAccept = true
 
-    const filePath = './WATCHER-test-files/WATCHER/xccdf/GoodStatistics-xccdf.xml'
+    const filePath = './WATCHER-test-files/WATCHER/cklb/GoodStatistics.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -242,6 +226,7 @@ describe('Testing that the xccdf Review Parser will return the correct figures i
   })
 
   it(' unreviewed: never, unreviewedCommented: informational, has comments/detail', async () => {
+    // will never import unreviewed findings
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'never',
@@ -264,7 +249,7 @@ describe('Testing that the xccdf Review Parser will return the correct figures i
 
     const allowAccept = true
 
-    const filePath = './WATCHER-test-files/WATCHER/xccdf/GoodStatistics-xccdf.xml'
+    const filePath = './WATCHER-test-files/WATCHER/cklb/GoodStatistics.cklb'
 
     const review = await generateReviewObject(
       filePath,
@@ -289,6 +274,7 @@ describe('Testing that the xccdf Review Parser will return the correct figures i
   })
 
   it(' unreviewed: never, unreviewedCommented: notchecked', async () => {
+    // will never import unreviewed findings
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'never',
@@ -311,7 +297,7 @@ describe('Testing that the xccdf Review Parser will return the correct figures i
 
     const allowAccept = true
 
-    const filePath = './WATCHER-test-files/WATCHER/xccdf/GoodStatistics-xccdf.xml'
+    const filePath = './WATCHER-test-files/WATCHER/cklb/GoodStatistics.cklb'
 
     const review = await generateReviewObject(
       filePath,
