+++ test/xccdf-tests/XCCDFReviewParserReviewObject.test.js
@@ -1,11 +1,9 @@
 import chai from 'chai';
-import { reviewsFromCkl } from '../ReviewParser.js'; 
+import { reviewsFromXccdf } from '../../ReviewParser.js'; 
 import { XMLParser } from 'fast-xml-parser';
 import fs from 'fs/promises';
 import he from 'he';
-
-const expect = chai.expect;
-
+const expect = chai.expect
 const valueProcessor = function (
   tagName,
   tagValue,
@@ -16,6 +14,25 @@ const valueProcessor = function (
   he.decode(tagValue)
 }
 
+const dataArray = [
+  {
+    scapBenchmarkId: 'CAN_Ubuntu_18-04_STIG',
+    benchmarkId: 'U_CAN_Ubuntu_18-04_STIG'
+  },
+  { scapBenchmarkId: 'Mozilla_Firefox_RHEL', benchmarkId: 'Mozilla_Firefox' },
+  {
+    scapBenchmarkId: 'Mozilla_Firefox_Windows',
+    benchmarkId: 'Mozilla_Firefox'
+  },
+  { scapBenchmarkId: 'MOZ_Firefox_Linux', benchmarkId: 'MOZ_Firefox_STIG' },
+  { scapBenchmarkId: 'MOZ_Firefox_Windows', benchmarkId: 'MOZ_Firefox_STIG' },
+  { scapBenchmarkId: 'Solaris_10_X86_STIG', benchmarkId: 'Solaris_10_X86' }
+]
+
+const scapBenchmarkMap = new Map(
+  dataArray.map(item => [item.scapBenchmarkId, item])
+)
+
 // Create a helper function to read the file and generate the review object
 async function generateReviewObject (
   filePath,
@@ -24,35 +41,31 @@ async function generateReviewObject (
   allowAccept
 ) {
   const data = await fs.readFile(filePath, 'utf8')
-  return reviewsFromCkl({
+  return reviewsFromXccdf({
     data,
-    importOptions,
     fieldSettings,
     allowAccept,
+    importOptions,
     valueProcessor,
+    scapBenchmarkMap,
     XMLParser
   })
 }
 
-
-describe('Import Options, allowAccept for a CKL review object in non multi-stig', () => {
+describe('Testing permutations of Import Options for a review objects parsed by xccdf Review Parser.', () => {
   it('DEFAULT SETTINGS: Primarily testing review "status = saved"', async () => {
     // Test: DEFAULT SETTINGS
-    // This test validates the behavior of the ckl parser function under default settings.
+    // This test validates the behavior of the xccdf parser function under default settings.
     // Primary Focus:
-    // - Ensuring that the 'autoStatus' option in 'importOptions' correctly sets the review status to 'saved'.
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // These checks ensure that not only is the 'status' property set as expected, but also that
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-notReviewed-Commented-Detailed.ckl') to simulate a real-world scenario.
+    // - Ensuring that the 'autoStatus' option in 'importOptions' correctly sets the reviews to be empty
+  
 
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
       unreviewedCommented: 'informational',
       emptyDetail: 'replace',
-      emptyComment: 'ignore',
+      emptyComment: 'replace',
       allowCustom: true
     }
 
@@ -70,8 +83,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-notReviewed-Commented-Detailed.ckl'
-
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-notReviewed-Commented-Detailed-xccdf.xml'
     const review = await generateReviewObject(
       filePath,
       importOptions,
@@ -79,26 +91,15 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
       allowAccept
     )
 
-    const expectedReview = {
-      ruleId: 'SV-257777r925318_rule',
-      status: 'saved',
-      result: 'informational',
-      comment: 'xyz',
-      detail: 'xyz'
-    }
-
-    expect(review.checklists[0].reviews[0]).to.include(expectedReview)
+    expect(review.checklists[0].reviews).to.be.empty
   })
 
   it('autoStatus = null (keep exisiting), testing that all review statuses do not exist', async () => {
     // Test: autoStatus = null
-    // This test validates the behavior of the ckl parser function under the 'autoStatus = null' setting.
+    // This test validates the behavior of the xccdf parser function under the 'autoStatus = null' setting.
     // Primary Focus:
-    // - Ensuring that the 'autoStatus' option in 'importOptions' correctly sets the review status to null (doesnt exist).
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'ruleId', 'result', 'comment', and 'detail'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-notReviewed-Commented-Detailed.ckl') to simulate a real-world scenario.
+    // - Ensuring that the 'autoStatus' option in 'importOptions' correctly sets the review to be empty
+  
     const importOptions = {
       autoStatus: 'null',
       unreviewed: 'commented',
@@ -122,7 +123,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-notReviewed-Commented-Detailed.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-notReviewed-Commented-Detailed-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -131,16 +132,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
       allowAccept
     )
 
-    const expectedReview = {
-      ruleId: 'SV-257777r925318_rule',
-      result: 'informational',
-      comment: 'xyz',
-      detail: 'xyz'
-    }
-
-    expect(review.checklists[0].reviews[0]).to.include(expectedReview)
-
-    expect(review.checklists[0].reviews[0].status).to.not.exist
+    expect(review.checklists[0].reviews).to.be.empty
   })
 
   it('autoStatus = submitted, testing if reviews are set to "submitted" if valid or "saved" if not valid. Determined by field settings and result', async () => {
@@ -148,11 +140,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     // This test validates the behavior of the ckl parser function under the autoStatus = submitted setting coupled with out field settings.
     // Primary Focus:
     // - Ensuring that the 'autoStatus' option in 'importOptions' correctly sets the review status to 'submitted' if valid or 'saved' if not valid.
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including ruleId 'result', 'comment', and 'detail'.
-    // These checks ensure that not only is the 'status' property set as expected, but also that
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-notReviewed-Commented-Detailed.ckl') to simulate a real-world scenario.
+    
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -176,7 +164,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Asset_a-VPN_TRUNCATED-V2R5.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Asset_a-VPN_TRUNCATED-V2R5-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -231,16 +219,12 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     }
   })
 
-  it('autoStatus = submitted, testing that we will set status to saved if does not meet field settings requirements', async () => {
+  it('autoStatus = submitted, testing that status is set to saved if does not meet field settings requirements', async () => {
     // Test: autoStatus = submitted
-    // This test validates the behavior of the ckl parser function under the autoStatus = submitted setting coupled with our field settings.
+    // This test validates the behavior of the xccdf parser function under the autoStatus = submitted setting coupled with our field settings.
     // Primary Focus:
-    // - Ensuring that the 'autoStatus' option in 'importOptions' correctly sets the review status to 'saved' if not valid.
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail'.
-    // These checks ensure that not only is the 'status' property set as expected, but also that
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-notReviewed-Commented-Detailed.ckl') to simulate a real-world scenario.
+    // - Ensuring that the 'autoStatus' option in 'importOptions' correctly sets the review to be empty
+
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -264,7 +248,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-notReviewed-Commented-Detailed.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-notReviewed-Commented-Detailed-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -272,29 +256,19 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
       fieldSettings,
       allowAccept
     )
-    const expectedReview = {
-      ruleId: 'SV-257777r925318_rule',
-      status: 'saved',
-      result: 'informational',
-      comment: 'xyz',
-      detail: 'xyz'
-    }
-    expect(review.checklists[0].reviews[0]).to.include(expectedReview)
+
+    expect(review.checklists[0].reviews).to.be.empty
   })
 
   it('(autoStatus = accepted, allowAccept = true) for permissions to use accepted status', async () => {
     // Test: autoStatus = accepted, allowAccept = true
-    // This test validates the behavior of the ckl parser function under the autoStatus = accepted, allowAccept = true settings.
+    // This test validates the behavior of the xccdf parser function under the autoStatus = accepted, allowAccept = true settings.
     // Primary Focus:
     // - Ensuring that the 'autoStatus' option in 'importOptions' correctly sets the review status to 'submitted' if the user does not have permissions
     // or 'accepted' if the user does have permissions (permissions are determined by the 'allowAccept' option).
     // note: if not accepted or submitted, it will be saved because review did not meet the field settings requirements
     // note: in this test we will have permissions to accept reviews
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // These checks ensure that not only is the 'status' property set as expected, but also that
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Asset_a-VPN_TRUNCATED-V2R5.ckl'') to simulate a real-world scenario.
+
 
     const importOptions = {
       autoStatus: 'accepted',
@@ -318,7 +292,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
 
     const allowAccept = true
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Asset_a-VPN_TRUNCATED-V2R5.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Asset_a-VPN_TRUNCATED-V2R5-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -375,19 +349,16 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
       expect(reviewItem.detail).to.equal(expectedDetail)
     }
   })
+
   it('(autoStatus = accepted allowAccept = false) for permissions to use submitted status', async () => {
     // Test: autoStatus = accepted, allowAccept = false
-    // This test validates the behavior of the ckl parser function under the autoStatus = accepted, allowAccept = false settings.
+    // This test validates the behavior of the xccdf parser function under the autoStatus = accepted, allowAccept = false settings.
     // Primary Focus:
     // - Ensuring that the 'autoStatus' option in 'importOptions' correctly sets the review status to 'submitted' if the user does not have permissions
     // or 'accepted' if the user does have permissions (permissions are determined by the 'allowAccept' option).
     // note: if not accepted or submitted, it will be saved because review did not meet the field settings requirements
     // note: in this test we will NOT have permissions to accept reviews (we should see submitted)
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // These checks ensure that not only is the 'status' property set as expected, but also that
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Asset_a-VPN_TRUNCATED-V2R5.ckl'') to simulate a real-world scenario.
+   
     const importOptions = {
       autoStatus: 'accepted',
       unreviewed: 'commented',
@@ -410,7 +381,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
 
     const allowAccept = false
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Asset_a-VPN_TRUNCATED-V2R5.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Asset_a-VPN_TRUNCATED-V2R5-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -467,16 +438,13 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     }
   })
 
-  it("'unreviewed = commented' testing that we only import unreviwred rules that contain a comment or detail   ", async () => {
+  it("'unreviewed = commented' testing that we ignore unreviewed reviews even if they have commment/detail  ", async () => {
     // Test: unreviewed = commented
-    // This test validates the behavior of the ckl parser function under the unreviewed = commented settings.
+    // This test validates the behavior of the xccdf parser function under the unreviewed = commented settings.
     // Primary Focus:
-    // - Ensuring that the 'unreviewed' option in 'importOptions' correctly imports only reviews with a non complience result that contain a comment and or detail.
+    // - Ensuring that the 'unreviewed' option in 'importOptions' correctly ignores reviews that are unreviewed and have a comment or detail.
     // note this test will have a review with a comment and detail
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-notReviewed-Commented-Detailed.ckl') to simulate a real-world scenario.
+
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -500,7 +468,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-notReviewed-Commented-Detailed.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-notReviewed-Commented-Detailed-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -509,33 +477,16 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
       allowAccept
     )
 
-    const expectedReview = {
-      ruleId: 'SV-257777r925318_rule',
-      status: 'saved',
-      result: 'informational',
-      comment: 'xyz',
-      detail: 'xyz'
-    }
-    // check that review has a comment or detail and exists
-    expect(review.checklists[0].reviews[0]).to.exist
-    expect(
-      review.checklists[0].reviews[0].detail ||
-        review.checklists[0].reviews[0].comment
-    ).to.not.be.null
-    // secondary check that review matches expected review object
-    expect(review.checklists[0].reviews[0]).to.include(expectedReview)
+    expect(review.checklists[0].reviews).to.be.empty
   })
 
-  it("'unreviewed = commented' testing that we only import unreviwred rules that contain a comment or detail but giving a review without either", async () => {
+  it("'unreviewed = commented' testing that we do not import unreviwred rules that contain a comment or detail but giving a review without either", async () => {
     // Test: unreviewed = commented
-    // This test validates the behavior of the ckl parser function under the unreviewed = commented settings.
+    // This test validates the behavior of the xccdf parser function under the unreviewed = commented settings.
     // Primary Focus:
     // - Ensuring that the 'unreviewed' option in 'importOptions' correctly imports only reviews with a non complience result that contain a comment and or detail.
     // note this test will not have any reviews because there will be no comment/detail
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-notReviewed-Commented-Detailed.ckl') to simulate a real-world scenario.
+    
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -559,7 +510,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-notReviewed-Empty-CommentDetail.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-notReviewed-Empty-CommentDetail-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -567,21 +518,18 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
       fieldSettings,
       allowAccept
     )
-      
+
     // check that review has a comment or detail and exisits
     expect(review.checklists[0].reviews).to.be.empty
   })
 
-  it("'unreviewed = always', testing that unreviewed always are import  ", async () => {
+  it("'unreviewed = always', testing that unreviewed always are import and result is 'notchecked'  ", async () => {
     // Test: unreviewed = always
-    // This test validates the behavior of the ckl parser function under the unreviewed = always settings.
+    // This test validates the behavior of the xccdf parser function under the unreviewed = always settings.
     // Primary Focus:
     // - Ensuring that the 'unreviewed' option in 'importOptions' correctly imports all reviews without a compliance result and if they have a comment
-    // or detail they will be labeled as informational.
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-notReviewed-Commented-Detailed.ckl') to simulate a real-world scenario.
+    // or detail they will be labeled as notchecked.
+
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'always',
@@ -605,7 +553,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-notReviewed-Commented-Detailed.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-notReviewed-Commented-Detailed-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -615,25 +563,23 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     )
 
     const expectedReview = {
-      ruleId: 'SV-257777r925318_rule',
+      ruleId: 'SV-2',
       status: 'saved',
-      result: 'informational',
+      result: 'notchecked',
       comment: 'xyz',
       detail: 'xyz'
     }
 
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
-  it("'unreviewed = always', testing that unreviewed always are import  ", async () => {
+
+  it("'unreviewed = always', testing that unreviewed always are imported ", async () => {
     // Test: unreviewed = always
-    // This test validates the behavior of the ckl parser function under the unreviewed = always settings.
+    // This test validates the behavior of the xccdf parser function under the unreviewed = always settings.
     // Primary Focus:
     // - Ensuring that the 'unreviewed' option in 'importOptions' correctly imports all reviews without a compliance result
     // and without comment or detail to be labled as notchecked.
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // These checks ensure that not only is the 'status' property set as expected, but also that
-    // The test utilizes a sample CKL file ('Single-Vuln-notReviewed-Empty-CommentDetail.ckl') to simulate a real-world scenario.
+
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'always',
@@ -657,7 +603,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-notReviewed-Empty-CommentDetail.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-notReviewed-Empty-CommentDetail-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -666,7 +612,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
       allowAccept
     )
     const expectedReview = {
-      ruleId: 'SV-257777r925318_rule',
+      ruleId: 'SV-2',
       status: 'saved',
       result: 'notchecked',
       comment: null,
@@ -675,15 +621,13 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
 
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
+
   it(" 'unreviewed = never' testing to never import an unreviewed item ", async () => {
     // Test: unreviewed = never
-    // This test validates the behavior of the ckl parser function under the unreviewed = never settings.
+    // This test validates the behavior of the xccdf parser function under the unreviewed = never settings.
     // Primary Focus:
     // - Ensuring that the 'unreviewed' option in 'importOptions' ignores reviews without a compliance result (Nf/na/o)
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-notReviewed-Empty-CommentDetail.ckl') to simulate a real-world scenario.
+    
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'never',
@@ -707,7 +651,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-notReviewed-Empty-CommentDetail.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-notReviewed-Empty-CommentDetail-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -719,16 +663,12 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     expect(review.checklists[0].reviews).to.be.empty
   })
 
-  it("'unreviewedComment = informational' testing that an unreviewed item with a comment has a result of informational", async () => {
+  it("'unreviewedComment = informational' testing that an unreviewed item with a comment is an empty review ", async () => {
     // Test: unreviewedComment = informational
-    // This test validates the behavior of the ckl parser function under the unreviewedComment = informational settings.
+    // This test validates the behavior of the xccdf parser function under the unreviewedComment = informational settings.
     // Primary Focus:
-    // - Ensuring that the 'unreviewed' option in 'importOptions' correctly labels a review without a compliance result..
-    // and with a comment or detail as informational.
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-notReviewed-Commented-Detailed.ckl') to simulate a real-world scenario.
+    // - Ensuring that the 'unreviewed' option in 'importOptions' correctly does not import the review.
+    
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -751,7 +691,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-notReviewed-Commented-Detailed.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-notReviewed-Commented-Detailed-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -760,26 +700,15 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
       allowAccept
     )
 
-    const expectedReview = {
-      ruleId: 'SV-257777r925318_rule',
-      status: 'saved',
-      result: 'informational',
-      comment: 'xyz',
-      detail: 'xyz'
-    }
-
-    expect(review.checklists[0].reviews[0]).to.include(expectedReview)
+    expect(review.checklists[0].reviews).to.be.empty
   })
-  it(" 'unreviewedComment = notchecked'. testing that an unreviewed with a comment has a result of notchecked", async () => {
+
+  it(" 'unreviewedComment = notchecked'. testing that an unreviewed with a comment returns no reviews", async () => {
     // Test: unreviewedComment = notchecked
-    // This test validates the behavior of the ckl parser function under the unreviewedComment = notchecked settings.
+    // This test validates the behavior of the xccdf parser function under the unreviewedComment = notchecked settings.
     // Primary Focus:
-    // - Ensuring that the 'unreviewed' option in 'importOptions' correctly labels a review without a compliance result..
-    // and with a comment or detail as notchecked.
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-notReviewed-Commented-Detailed.ckl') to simulate a real-world scenario.
+    // - Ensuring that the 'unreviewed' option in 'importOptions' correctly ignores a review without a compliance result
+
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -802,7 +731,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-notReviewed-Commented-Detailed.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-notReviewed-Commented-Detailed-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -810,25 +739,15 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
       fieldSettings,
       allowAccept
     )
-    const expectedReview = {
-      ruleId: 'SV-257777r925318_rule',
-      status: 'saved',
-      result: 'notchecked',
-      comment: 'xyz',
-      detail: 'xyz'
-    }
 
-    expect(review.checklists[0].reviews[0]).to.include(expectedReview)
+    expect(review.checklists[0].reviews).to.be.empty
   })
+
   it("'emptyDetail = replace' testing that if an item has an empty detail we will replace it with a static message", async () => {
     // Test: emptyDetail = replace
-    // This test validates the behavior of the ckl parser function under the emptyDetail = replace settings.
+    // This test validates the behavior of the xccdf parser function under the emptyDetail = replace settings.
     // Primary Focus:
-    // - Ensuring that the 'emptyDetail' option in 'importOptions' correctly replaces an empty detail with a static message. "There is no detail provided for the assessment"
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-Pass-With-Comment.ckl') to simulate a real-world scenario.
+    // - Ensuring that the 'emptyDetail' option in 'importOptions' correctly replaces an empty detail with a static message.
 
     const importOptions = {
       autoStatus: 'saved',
@@ -852,7 +771,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-Pass-With-Comment.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-Pass-With-Comment-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -861,23 +780,21 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
       allowAccept
     )
     const expectedReview = {
-      ruleId: 'SV-207191r803418_rule',
+      ruleId: 'SV-2',
       status: 'saved',
       result: 'pass',
       comment: 'xyz',
-      detail: 'There is no detail provided for the assessment'
+      detail:
+        'Result was reported by product "MyTestSystem" version 1 at 2023-11-13T16:41:49.000Z using check content "MyCheckContent"'
     }
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
-  it("'emptyDetail = ignore' testing that if there is no detail provided it will retaing exisiting, if no exisitng then we will set to null", async () => {
+
+  it("'emptyDetail = ignore' testing that if there is no detail provided it will retain exisiting detail, if no exisitng then we will set to null", async () => {
     // Test: emptyDetail = ignore
-    // This test validates the behavior of the ckl parser function under the emptyDetail = ignore settings.
+    // This test validates the behavior of the xccdf parser function under the emptyDetail = ignore settings.
     // Primary Focus:
     // - Ensuring that the 'emptyDetail' option in 'importOptions' correctly replaces an empty detail a null value
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-Pass-With-Comment.ckl') to simulate a real-world scenario.
 
     const importOptions = {
       autoStatus: 'saved',
@@ -901,7 +818,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-Pass-With-Comment.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-Pass-With-Comment-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -911,7 +828,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803418_rule',
+      ruleId: 'SV-2',
       status: 'saved',
       result: 'pass',
       comment: 'xyz',
@@ -922,13 +839,10 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
   })
   it("'emptyDetail = import', testing empty detail will clear existing text (setting it to empty string)", async () => {
     // Test: emptyDetail = import
-    // This test validates the behavior of the ckl parser function under the emptyDetail = import settings.
+    // This test validates the behavior of the xccdf parser function under the emptyDetail = import settings.
     // Primary Focus:
     // - Ensuring that the 'emptyDetail' option in 'importOptions' correctly replaces an empty detail an empty string if no detail is provided
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-Pass-With-Comment.ckl') to simulate a real-world scenario.
+   
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -951,7 +865,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-Pass-With-Comment.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-Pass-With-Comment-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -961,7 +875,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803418_rule',
+      ruleId: 'SV-2',
       status: 'saved',
       result: 'pass',
       comment: 'xyz',
@@ -973,13 +887,10 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
 
   it(" 'emptyDetail = import' testing that a review with a detail provided will be applied", async () => {
     // Test: emptyDetail = import
-    // This test validates the behavior of the ckl parser function under the emptyDetail = import settings.
+    // This test validates the behavior of the xccdf parser function under the emptyDetail = import settings.
     // Primary Focus:
     // - Ensuring that the 'emptyDetail' option in 'importOptions' correctly uses the exisisitng detail if one is provided
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-Pass-With-Detail.ckl') to simulate a real-world scenario.
+  
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -1002,7 +913,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-Pass-With-Detail.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-Pass-With-Detail-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -1012,7 +923,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803418_rule',
+      ruleId: 'SV-2',
       status: 'saved',
       result: 'pass',
       comment: null,
@@ -1021,15 +932,12 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
 
-  it("'emptyComment = replace' testing that if an item has an empty comment we will replace it with a static message", async () => {
+  it("'emptyComment = replace' testing that if an item has an empty comment we will replace it with a dynamically created message", async () => {
     // Test: emptyComment = replace
-    // This test validates the behavior of the ckl parser function under the emptyComment = replace settings.
+    // This test validates the behavior of the xxcdf parser function under the emptyComment = replace settings.
     // Primary Focus:
-    // - Ensuring that the 'emptyComment' option in 'importOptions' correctly replaces an empty comment with a static message. "There is no comment provided for the assessment"
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-Pass-With-Detail.ckl') to simulate a real-world scenario.
+    // - Ensuring that the 'emptyComment' option in 'importOptions' correctly replaces an empty comment with a dynamically created message based off the data.
+   
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -1052,7 +960,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-Pass-With-Detail.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-Pass-With-Detail-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -1062,24 +970,23 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803418_rule',
+      ruleId: 'SV-2',
       status: 'saved',
       result: 'pass',
-      comment: 'There is no comment provided for the assessment',
+      comment:
+        'Result was reported by product "MyTestSystem" version 1 at 2023-11-13T16:41:49.000Z using check content "MyCheckContent"',
       detail: 'xyz'
     }
 
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
+
   it("'emptyComment = ignore' testing that we will use exisitng text, if none use null", async () => {
     // Test: emptyComment = ignore
-    // This test validates the behavior of the ckl parser function under the emptyComment = ignore settings.
+    // This test validates the behavior of the xccdf parser function under the emptyComment = ignore settings.
     // Primary Focus:
     // - Ensuring that the 'emptyComment' option in 'importOptions' correctly replaces an empty comment a null value
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-Pass-With-Comment.ckl') to simulate a real-world scenario.
+  
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -1102,7 +1009,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-Pass-With-Detail.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-Pass-With-Detail-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -1111,7 +1018,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
       allowAccept
     )
     const expectedReview = {
-      ruleId: 'SV-207191r803418_rule',
+      ruleId: 'SV-2',
       status: 'saved',
       result: 'pass',
       comment: null,
@@ -1119,15 +1026,13 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     }
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
+
   it("'emptyComment = import', will clear eixsitng with an empty string if no comment given ", async () => {
     // Test: emptyComment = import
-    // This test validates the behavior of the ckl parser function under the emptyComment = import settings.
+    // This test validates the behavior of the xccdf parser function under the emptyComment = import settings.
     // Primary Focus:
     // - Ensuring that the 'emptyComment' option in 'importOptions' correctly replaces an empty comment an empty string: ""
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-Pass-With-Detail.ckl')
+    
 
     const importOptions = {
       autoStatus: 'saved',
@@ -1151,7 +1056,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-Pass-With-Detail.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-Pass-With-Detail-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -1161,7 +1066,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803418_rule',
+      ruleId: 'SV-2',
       status: 'saved',
       result: 'pass',
       comment: '',
@@ -1169,15 +1074,13 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     }
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
+
   it("'emptyComment = import', testing a review with a comment provided in the ckl to make sure we get it back in the review ", async () => {
     // Test: emptyComment = import
-    // This test validates the behavior of the ckl parser function under the emptyComment = import settings.
+    // This test validates the behavior of the xccdf parser function under the emptyComment = import settings.
     // Primary Focus:
     // - Ensuring that the 'emptyComment' option in 'importOptions' correctly uses the exisisitng comment if one is provided
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-Pass-With-Detail.ckl')
+ 
     const importOptions = {
       autoStatus: 'saved',
       unreviewed: 'commented',
@@ -1200,7 +1103,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-Pass-With-Comment.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-Pass-With-Comment-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -1210,7 +1113,7 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803418_rule',
+      ruleId: 'SV-2',
       status: 'saved',
       result: 'pass',
       comment: 'xyz',
@@ -1220,16 +1123,13 @@ describe('Import Options, allowAccept for a CKL review object in non multi-stig'
   })
 })
 
-describe('fieldSettings testing for a CKL review object in non multi-stig', () => {
+describe('fieldSettings testing for a review object in non multi-stig xccdf', () => {
   it("DEFAULT FIELD SETTINGS with allowAccept=true and a passing review, testing that it has a detail and is 'submitted'", async () => {
     // Test: autostatus = submitted, default field settings.
-    // This test validates the behavior of the ckl parser function under the default field settings.
+    // This test validates the behavior of the xccdf parser function under the default field settings.
     // Primary Focus:
     // - Ensuring that the 'autoStatus' option in 'importOptions' correctly ensures that a detail is required for a review to be submitted.
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-Pass-With-Detail.ckl')
+   
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1253,7 +1153,7 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-Pass-With-Detail.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-Pass-With-Detail-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -1263,7 +1163,7 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803418_rule',
+      ruleId: 'SV-2',
       status: 'submitted',
       result: 'pass',
       comment: '',
@@ -1273,16 +1173,14 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
     // expected status is submitted for the rule that has a detail
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
+
   it('DEFAULT FIELD SETTINGS  with allowAccept=true and a failing review with no detail.', async () => {
     //autostatus = submitted, testing default field settings with allowAccept=true
-    // This test validates the behavior of the ckl parser function under the default field setting with a fail and a no detail.
+    // This test validates the behavior of the xccdf parser function under the default field setting with a fail and a no detail.
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a detail is always required for submission for a review.
     // Test that with a failing review and no detail exisiting and it will be set to 'saved'
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-fail-With-Detail.ckl')
+   
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1306,7 +1204,7 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-fail-Empty-CommentDetail.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-fail-Empty-CommentDetail-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -1316,7 +1214,7 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803418_rule',
+      ruleId: 'SV-2',
       status: 'saved',
       result: 'fail',
       comment: null,
@@ -1325,16 +1223,14 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
 
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
+
   it('fieldSettings.detail.required = findings with allowAccept=true with a failing review containing a detail', async () => {
     // Test: autostatus = submitted, default field settings.
-    // This test validates the behavior of the ckl parser function under the  fieldSettings.detail.required = findings field settings.
+    // This test validates the behavior of the xccdf parser function under the  fieldSettings.detail.required = findings field settings.
     // Primary Focus:
     // - Ensuring that the 'autoStatus' option in 'importOptions' correctly ensures that a detail is required for a review  that has findings to be submitted.
     // Test that a failing review with a detail will be submitted
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-Pass-With-Detail.ckl')
+   
 
     const importOptions = {
       autoStatus: 'submitted',
@@ -1359,7 +1255,7 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-fail-With-Detail.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-fail-With-Detail-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -1369,7 +1265,7 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803418_rule',
+      ruleId: 'SV-2',
       status: 'submitted',
       result: 'fail',
       comment: '',
@@ -1378,16 +1274,14 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
 
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
+
   it("'fieldSettings.detail.required = findings' with allowAccept=true with a fail and no detail or comment", async () => {
     // Test: autostatus = submitted, testing 'fieldSettings.detail.required = findings' allowAccept=true
-    // This test validates the behavior of the ckl parser function under the  fieldSettings.detail.required = findings field setting with a fail and no detail or comment.
+    // This test validates the behavior of the xccdf parser function under the  fieldSettings.detail.required = findings field setting with a fail and no detail or comment.
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a detail is required for a review that has findings to be submitted if not we will save.
     // Test that no detail exisitng and it will be set to 'saved'
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-fail-Empty-CommentDetail.ckl')
+   
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1411,7 +1305,7 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-fail-Empty-CommentDetail.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-fail-Empty-CommentDetail-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -1420,28 +1314,23 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
       allowAccept
     )
     const expectedReview = {
-      ruleId: 'SV-207191r803418_rule',
+      ruleId: 'SV-2',
       status: 'saved',
       result: 'fail',
       comment: null,
       detail: null
     }
 
-    // console.log(JSON.stringify(review, null, 2))
-
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
 
   it("'fieldSettings.detail.required = optional' with allowAccept=true with a fail and no detail or comment, testing that it does not have a detail and is submitted ", async () => {
     // TEST: autostatus = submitted, testing 'fieldSettings.detail.required = optional' with allowAccept=true
-    // This test validates the behavior of the ckl parser function under the  fieldSettings.detail.required = optional field setting with a fail and no detail or comment.
+    // This test validates the behavior of the xccdf parser function under the  fieldSettings.detail.required = optional field setting with a fail and no detail or comment.
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a detail is optional for submission for a review that has findings.
     // Test that no detail exisitng and it will be set to 'submitted'
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-fail-Empty-CommentDetail.ckl')
+   
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1465,7 +1354,7 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-fail-Empty-CommentDetail.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-fail-Empty-CommentDetail-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -1475,7 +1364,7 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803418_rule',
+      ruleId: 'SV-2',
       status: 'submitted',
       result: 'fail',
       comment: null,
@@ -1484,16 +1373,14 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
 
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
+
   it("fieldSettings.detail.required = optional' with allowAccept=true with a fail and detail testing it has a detail and is submitted", async () => {
     // autostatus = submitted, 'fieldSettings.detail.required = optional' with allowAccept=true
-    // This test validates the behavior of the ckl parser function under the  fieldSettings.detail.required = optional field setting with a fail and a detail.
+    // This test validates the behavior of the xccdf parser function under the  fieldSettings.detail.required = optional field setting with a fail and a detail.
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a detail is optional for submission for a review that has findings.
     // Test that with detail exisitng and it will be set to 'submitted'
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-fail-With-Detail.ckl')
+   
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1517,7 +1404,7 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-fail-With-Detail.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-fail-With-Detail-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -1527,7 +1414,7 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803418_rule',
+      ruleId: 'SV-2',
       status: 'submitted',
       result: 'fail',
       comment: null,
@@ -1539,14 +1426,11 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
 
   it('DEFAULT FIELD SETTINGS with allowAccept=true and a passing review testing that it has a comment and is submitted', async () => {
     //autostatus = submitted, testing default field settings with allowAccept=true
-    // This test validates the behavior of the ckl parser function under the default field setting with a pass and a comment.
+    // This test validates the behavior of the xccdf parser function under the default field setting with a pass and a comment.
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a comment is always required for submission for a review.
     // Test that witha passing review and a comment exisiting and it will be set to 'submitted'
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-fail-With-Detail.ckl')
+   
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1570,7 +1454,7 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-Pass-With-Comment.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-Pass-With-Comment-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -1580,7 +1464,7 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803418_rule',
+      ruleId: 'SV-2',
       status: 'submitted',
       result: 'pass',
       comment: 'xyz',
@@ -1589,16 +1473,14 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
 
     expect(review.checklists[0].reviews[0]).to.include(expectedReview)
   })
+
   it('DEFAULT FIELD SETTINGS  with allowAccept=true and a failing review with no comment.', async () => {
     //autostatus = submitted, testing default field settings with allowAccept=true
-    // This test validates the behavior of the ckl parser function under the default field setting with a fail and a no comment.
+    // This test validates the behavior of the xccdf parser function under the default field setting with a fail and a no comment.
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a comment is always required for submission for a review.
     // Test that with a failing review and no comment exisiting and it will be set to 'saved'
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-fail-With-Detail.ckl')
+    
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1622,7 +1504,7 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-fail-Empty-CommentDetail.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-fail-Empty-CommentDetail-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -1632,7 +1514,7 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803418_rule',
+      ruleId: 'SV-2',
       status: 'saved',
       result: 'fail',
       comment: null,
@@ -1643,14 +1525,11 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
   })
   it('fieldSettings.comment.required = findings with allowAccept=true with a fail and comment testing that it has a comment and is submitted', async () => {
     // TEST: autostatus = submitted, fieldSettings.comment.required = findings with allowAccept=true
-    // This test validates the behavior of the ckl parser function under the fieldSettings.comment.required = findings field setting with a fail and a comment.
+    // This test validates the behavior of the xccdf parser function under the fieldSettings.comment.required = findings field setting with a fail and a comment.
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a comment is always required for submission for a review that contains a finding.
     // Test that with a failed review and a comment exisitng and it will be set to 'submitted'
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-fail-with-Comment.ckl')
+    
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1674,7 +1553,7 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-fail-with-Comment.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-fail-with-Comment-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -1684,7 +1563,7 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803418_rule',
+      ruleId: 'SV-2',
       status: 'submitted',
       result: 'fail',
       comment: 'xyz',
@@ -1695,14 +1574,11 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
   })
   it("'fieldSettings.comment.required = findings' with allowAccept=true with a fail and no detail or comment, testing that it does not have a comment and is 'saved' ", async () => {
     // TEST: autostatus = submitted, testing 'fieldSettings.comment.required = findings' with allowAccept=true
-    // This test validates the behavior of the ckl parser function under the fieldSettings.comment.required = findings field setting with a fail and no comment.
+    // This test validates the behavior of the xccdf parser function under the fieldSettings.comment.required = findings field setting with a fail and no comment.
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a comment is always required for submission for a review that contains a finding.
     // Test that with a failed review and no comment will be set to 'saved'
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-fail-Empty-CommentDetail.ckl')
+    
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1726,7 +1602,7 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-fail-Empty-CommentDetail.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-fail-Empty-CommentDetail-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -1736,7 +1612,7 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803418_rule',
+      ruleId: 'SV-2',
       status: 'saved',
       result: 'fail',
       comment: null,
@@ -1748,14 +1624,11 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
 
   it("'fieldSettings.comment.required = optional' with allowAccept=true with a fail and no detail or comment. testing that it doesnt have a comment and is submmited", async () => {
     // TEST: autostatus = submitted, testing 'fieldSettings.comment.required = optional' with allowAccept=true
-    // This test validates the behavior of the ckl parser function under the fieldSettings.comment.required = optional field setting with a fail and no comment or detail .
+    // This test validates the behavior of the xccdf parser function under the fieldSettings.comment.required = optional field setting with a fail and no comment or detail .
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a comment is optionally required for submission for a review that contains a finding.
     // Test that with a failed review and no comment will be set to 'submitted'
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-fail-Empty-CommentDetail.ckl')
+ 
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1779,7 +1652,7 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-fail-Empty-CommentDetail.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-fail-Empty-CommentDetail-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -1789,7 +1662,7 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803418_rule',
+      ruleId: 'SV-2',
       status: 'submitted',
       result: 'fail',
       comment: null,
@@ -1800,14 +1673,11 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
   })
   it("fieldSettings.comment.required = optional' with allowAccept=true with a fail and comment, testing thhat it has a comment and is submitted", async () => {
     // TEST: autostatus = submitted, 'fieldSettings.comment.required = optional' with allowAccept=true
-    // This test validates the behavior of the ckl parser function under the fieldSettings.comment.required = optional field setting with a fail and a comment.
+    // This test validates the behavior of the xccdf parser function under the fieldSettings.comment.required = optional field setting with a fail and a comment.
     // Primary Focus:
     // - ensuring that fieldSettings correctly ensures that a comment is optionally required for submission for a review that contains a finding.
     // Test that with a failed review and a comment will be set to 'submitted'
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-fail-with-Comment.ckl')
+
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1831,7 +1701,7 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-fail-with-Comment.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-fail-with-Comment-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -1841,7 +1711,7 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
     )
 
     const expectedReview = {
-      ruleId: 'SV-207191r803418_rule',
+      ruleId: 'SV-2',
       status: 'submitted',
       result: 'fail',
       comment: 'xyz',
@@ -1852,17 +1722,14 @@ describe('fieldSettings testing for a CKL review object in non multi-stig', () =
   })
 })
 
-describe(' CKL Tests where fieldSettings and importOptions overlap. ', () => {
+describe('Tests where fieldSettings and importOptions overlap xccdf. ', () => {
   it("Testing where emptyDetail: 'ignore', emptyComment: 'ignore', aswell as requiring a comment and detail  ", async () => {
     // TEST: emptyDetail: 'ignore', emptyComment: 'ignore', fieldSettings.detail.required = always, fieldSettings.comment.required = always
-    // This test validates the behavior of the ckl parser function under above settings with a non compliance resilt and no comment or detail.
+    // This test validates the behavior of the xccdf parser function under above settings with a non compliance resilt and no comment or detail.
     // Primary Focus:
     // - ensuring that we will have no reviews because a brand new review will be created with "null" comment or detail with are both required.
     // Test that with a failed review and a comment will be set to 'submitted'
-    // Secondary Focus:
-    // - Verifying the accuracy of other review properties including 'result', 'comment', and 'detail', 'ruleId'.
-    // other related properties in the review object reflect the correct values as per the given 'importOptions' and 'fieldSettings'.
-    // The test utilizes a sample CKL file ('Single-Vuln-fail-with-Comment.ckl')
+   
     const importOptions = {
       autoStatus: 'submitted',
       unreviewed: 'commented',
@@ -1886,7 +1753,7 @@ describe(' CKL Tests where fieldSettings and importOptions overlap. ', () => {
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-notReviewed-Empty-CommentDetail.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-notReviewed-Empty-CommentDetail-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -1899,14 +1766,12 @@ describe(' CKL Tests where fieldSettings and importOptions overlap. ', () => {
   })
 })
 
-describe('MISC CKL. ', () => {
-  it('Testing that long comment.detail is truncated ', async () => {
-    // NOTE: the input comment and detail are '32768' characters long
-
+describe('MISC. xccdf ', () => {
+  it('review with result engine data', async () => {
     const importOptions = {
-      autoStatus: 'submitted',
-      unreviewed: 'commented',
-      unreviewedCommented: 'informational',
+      autoStatus: 'saved',
+      unreviewed: 'never',
+      unreviewedCommented: 'notchecked',
       emptyDetail: 'ignore',
       emptyComment: 'ignore',
       allowCustom: true
@@ -1914,19 +1779,19 @@ describe('MISC CKL. ', () => {
 
     const fieldSettings = {
       detail: {
-        enabled: 'findings', // not used
+        enabled: 'always',
         required: 'always'
       },
       comment: {
-        enabled: 'always', // not used
-        required: 'always'
+        enabled: 'findings',
+        required: 'findings'
       }
     }
 
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/Single-Vuln-Long-CommentDetail.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/Single-Vuln-Pass-With-Comment-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -1935,72 +1800,35 @@ describe('MISC CKL. ', () => {
       allowAccept
     )
 
-    const maxLength = 32767
+    //  console.log(JSON.stringify(review, null, 2))
 
-    expect(review.checklists[0].reviews[0].detail).to.have.lengthOf(maxLength)
-    expect(review.checklists[0].reviews[0].comment).to.have.lengthOf(maxLength)
-  })
-  it('Testing result engine ckl with an expression of the Eval STIG "module" that did the evaluation  ', async () => {
-    const importOptions = {
-      autoStatus: 'submitted',
-      unreviewed: 'commented',
-      unreviewedCommented: 'informational',
-      emptyDetail: 'ignore',
-      emptyComment: 'ignore',
-      allowCustom: true
-    }
-
-    const fieldSettings = {
-      detail: {
-        enabled: 'findings', // not used
-        required: 'always'
-      },
-      comment: {
-        enabled: 'always', // not used
-        required: 'always'
-      }
-    }
-
-    const allowAccept = true
-
-    const filePath =
-      './WATCHER-test-files/WATCHER/ckl/ResultEngineWithEvalStigModuleAndOverride.ckl'
-
-    const review = await generateReviewObject(
-      filePath,
-      importOptions,
-      fieldSettings,
-      allowAccept
-    )
-
-    const expectedResultEngine = {
-      type: 'script',
-      product: 'Evaluate-STIG',
-      version: '1.2204.1',
-      time: '2022-06-03T12:19:27.9454169-04:00',
-      checkContent: {
-        location: 'Scan-Windows10_Checks'
-      },
-      overrides: [
-        {
-          authority: 'MS_Windows_10_STIG_Answer_file.xml',
-          oldResult: 'unknown',
-          newResult: 'notapplicable',
-          remark: 'Evaluate-STIG Answer File'
+    const expectedResultEngineReview = {
+      ruleId: 'SV-2',
+      result: 'pass',
+      resultEngine: {
+        time: '2023-11-13T16:41:49.000Z',
+        type: 'scap',
+        version: "1",
+        product: 'MyTestSystem',
+        checkContent: {
+          location: 'MyCheckContent',
+          component: 'fso_comp_MyStig:135'
         }
-      ]
+      },
+      detail: null,
+      comment: 'xyz',
+      status: 'saved'
     }
 
-    expect(review.checklists[0].reviews[0].resultEngine).to.deep.equal(
-      expectedResultEngine
+    expect(review.checklists[0].reviews[0]).to.deep.equal(
+      expectedResultEngineReview
     )
   })
-  it('Testing result engine ckl with an  Eval STIG individual answer file override', async () => {
-
+  it('review with result override', async () => {
     const importOptions = {
-      autoStatus: 'submitted',
-      unreviewed: 'commented',
-      unreviewedCommented: 'informational',
+      autoStatus: 'saved',
+      unreviewed: 'never',
+      unreviewedCommented: 'notchecked',
       emptyDetail: 'ignore',
       emptyComment: 'ignore',
       allowCustom: true
@@ -2008,19 +1836,19 @@ describe('MISC CKL. ', () => {
 
     const fieldSettings = {
       detail: {
-        enabled: 'findings', // not used
+        enabled: 'always',
         required: 'always'
       },
       comment: {
-        enabled: 'always', // not used
-        required: 'always'
+        enabled: 'findings',
+        required: 'findings'
       }
     }
 
     const allowAccept = true
 
     const filePath =
-      './WATCHER-test-files/WATCHER/ckl/ResultEngineWithEvalStigModule.ckl'
+      './WATCHER-test-files/WATCHER/xccdf/ReviewOverrides-xccdf.xml'
 
     const review = await generateReviewObject(
       filePath,
@@ -2029,123 +1857,16 @@ describe('MISC CKL. ', () => {
       allowAccept
     )
 
-    const expectedResultEngine = {
-      type: 'script',
-      product: 'Evaluate-STIG',
-      version: '1.2204.1',
-      time: '2022-06-03T12:19:27.9454169-04:00',
-      checkContent: {
-        location: 'Scan-Windows10_Checks'
-      }
+
+    const expectedOveride = {
+      authority: 'Authority1',
+      oldResult: 'pass',
+      newResult: 'fail',
+      remark: 'Some remark'
     }
-    expect(review.checklists[0].reviews[0].resultEngine).to.deep.equal(
-      expectedResultEngine
+
+    expect(review.checklists[0].reviews[0].resultEngine.overrides[0]).to.deep.equal(
+      expectedOveride
     )
-  })
-  it('Testing result engine ckl with an Eval STIG individual answer file override that is incorrect', async () => {
-    const importOptions = {
-      autoStatus: 'submitted',
-      unreviewed: 'commented',
-      unreviewedCommented: 'informational',
-      emptyDetail: 'ignore',
-      emptyComment: 'ignore',
-      allowCustom: true
-    }
-
-    const fieldSettings = {
-      detail: {
-        enabled: 'findings', // not used
-        required: 'always'
-      },
-      comment: {
-        enabled: 'always', // not used
-        required: 'always'
-      }
-    }
-
-    const allowAccept = true
-
-    const filePath =
-      './WATCHER-test-files/WATCHER/ckl/ResultEngineInvalidComment.ckl'
-
-    const review = await generateReviewObject(
-      filePath,
-      importOptions,
-      fieldSettings,
-      allowAccept
-    )
-
-    expect(review.checklists[0].reviews[0].resultEngine).to.be.null
-  })
-  it('Testing no <STATUS> value in a ckl ', async () => {
-    // NOTE: the input comment and detail are '32768' characters long
-
-    const importOptions = {
-      autoStatus: 'submitted',
-      unreviewed: 'commented',
-      unreviewedCommented: 'informational',
-      emptyDetail: 'ignore',
-      emptyComment: 'ignore',
-      allowCustom: true
-    }
-
-    const fieldSettings = {
-      detail: {
-        enabled: 'findings', // not used
-        required: 'always'
-      },
-      comment: {
-        enabled: 'always', // not used
-        required: 'always'
-      }
-    }
-
-    const allowAccept = true
-
-    const filePath = './WATCHER-test-files/WATCHER/ckl/NoResult.ckl'
-
-    const review = await generateReviewObject(
-      filePath,
-      importOptions,
-      fieldSettings,
-      allowAccept
-    )
-
-    expect(review.checklists[0].reviews).to.be.empty
-  })
-  it('Testing no RuleID value in a ckl ', async () => {
-
-    const importOptions = {
-      autoStatus: 'submitted',
-      unreviewed: 'commented',
-      unreviewedCommented: 'informational',
-      emptyDetail: 'ignore',
-      emptyComment: 'ignore',
-      allowCustom: true
-    }
-
-    const fieldSettings = {
-      detail: {
-        enabled: 'findings', // not used
-        required: 'always'
-      },
-      comment: {
-        enabled: 'always', // not used
-        required: 'always'
-      }
-    }
-
-    const allowAccept = true
-
-    const filePath = './WATCHER-test-files/WATCHER/ckl/NoRuleId.ckl'
-
-    const review = await generateReviewObject(
-      filePath,
-      importOptions,
-      fieldSettings,
-      allowAccept
-    )
-
-    expect(review.checklists[0].reviews).to.be.empty
   })
 })
